### 2020-08-31 #2618 @whedon
[PRE REVIEW]: EMBERS: Experimental Measurement of BEam Responses with Satellites
**Submitting author:** @amanchokshi (<a href="http://orcid.org/0000-0003-1130-6390">Aman Chokshi</a>)
**Repository:** <a href="https://github.com/amanchokshi/EMBERS" target ="_blank">https://github.com/amanchokshi/EMBERS</a>
**Version:** v0.7.0
