### 2020-09-01 #2629 @mbobra
👋 @teuben @hasantahir Thank you so much for agreeing to review this submission! You'll see your reviewer checklists above. If you'd like more information on any of these items, please reference the [review critera](https://joss.readthedocs.io/en/latest/review_criteria.html). Also, we encourage open communication between the reviewers, submitting author, and editor. So please feel free to ask questions here on this review thread. I'm happy to help!
