### 2020-09-08 a8587ac
Script to compare corresponding slices of beam maps
