### 2020-09-09 4cf113e
Cleaned up data structures
