### 2020-09-09 c3fb8e1
Cleaned up data structures
