### 2020-09-13 949aad3
Preliminary tile gradient calculations
