### 2020-09-13 6df5802
Much better code to fit plane to 3D data
