### 2020-09-14 1aff3d8
Stupidly long list on healpix indices for primary beam
