+++ paper_plots/tile_tilt.py
@@ -12,6 +12,7 @@ import healpy as hp
 import matplotlib
 import numpy as np
 from embers.rf_tools.colormaps import jade
+from embers.tile_maps.beam_utils import plot_healpix
 from matplotlib import pyplot as plt
 from scipy.linalg import lstsq
 
@@ -86,7 +87,6 @@ def beam_maps(f):
     f_name, _ = Path(f).name.split(".")
     t_name, r_name, _, _ = f_name.split("_")
 
-    # pointings = ['0','2','4','41']
     pointings = ["0", "2", "4"]
 
     maps = []
@@ -108,16 +108,962 @@ def beam_maps(f):
         # healpix meadian map
         tile_med = np.asarray([(np.nanmedian(j) if j != [] else np.nan) for j in tile])
 
-        residuals = tile_med - fee
-        residuals[np.where(fee < -30)] = np.nan
-        residuals[np.where(tile_med == np.nan)] = np.nan
+        # Here ends the primary beam
+        pb_ind = np.where(fee < -80)[0][0]
 
-        maps.append([tile_med, residuals])
+        #  print(list(np.where(fee >= -30)[0]))
+
+        XX_0_PB = [
+            0,
+            1,
+            2,
+            3,
+            4,
+            5,
+            6,
+            7,
+            8,
+            9,
+            10,
+            11,
+            12,
+            13,
+            14,
+            15,
+            16,
+            17,
+            18,
+            19,
+            20,
+            21,
+            22,
+            23,
+            24,
+            25,
+            26,
+            27,
+            28,
+            29,
+            30,
+            31,
+            32,
+            33,
+            34,
+            35,
+            36,
+            37,
+            38,
+            39,
+            40,
+            41,
+            42,
+            43,
+            44,
+            45,
+            46,
+            47,
+            48,
+            49,
+            50,
+            51,
+            52,
+            53,
+            54,
+            55,
+            56,
+            57,
+            58,
+            59,
+            60,
+            61,
+            62,
+            63,
+            64,
+            65,
+            66,
+            67,
+            68,
+            69,
+            70,
+            71,
+            72,
+            73,
+            74,
+            75,
+            76,
+            77,
+            78,
+            79,
+            80,
+            81,
+            82,
+            83,
+            84,
+            85,
+            86,
+            87,
+            88,
+            89,
+            90,
+            91,
+            92,
+            93,
+            94,
+            95,
+            96,
+            97,
+            98,
+            99,
+            100,
+            101,
+            102,
+            103,
+            104,
+            105,
+            106,
+            107,
+            108,
+            109,
+            110,
+            111,
+            112,
+            113,
+            114,
+            115,
+            116,
+            117,
+            118,
+            119,
+            120,
+            121,
+            122,
+            123,
+            124,
+            125,
+            126,
+            127,
+            128,
+            129,
+            130,
+            131,
+            132,
+            133,
+            134,
+            135,
+            136,
+            137,
+            138,
+            139,
+            140,
+            141,
+            142,
+            143,
+            144,
+            145,
+            146,
+            147,
+            148,
+            149,
+            150,
+            151,
+            152,
+            153,
+            154,
+            155,
+            156,
+            157,
+            158,
+            159,
+            160,
+            161,
+            162,
+            163,
+            164,
+            165,
+            166,
+            167,
+            168,
+            169,
+            170,
+            171,
+            172,
+            173,
+            174,
+            175,
+            176,
+            177,
+            178,
+            179,
+            180,
+            181,
+            182,
+            183,
+            184,
+            185,
+            186,
+            187,
+            188,
+            189,
+            190,
+            191,
+            192,
+            193,
+            194,
+            195,
+            196,
+            197,
+            198,
+            199,
+            200,
[…]
