### 2020-09-14 7e984ab
Masks for 0, 2, 4
