### 2020-09-14 bccd2cd
Gradients calculated properly
