### 2020-09-16 f72df4b
Fixed Frequency x-lable typo
