### 2020-09-16 a5ef85d
Combined plot
