### 2020-09-18 efea8b5
Script to incestigate the beam errors along cardinal axes for both pols
