+++ paper_plots/cardinal_errors.py
@@ -1,0 +1,236 @@
+import argparse
+import concurrent.futures
+from itertools import repeat
+from pathlib import Path
+
+import matplotlib
+import numpy as np
+from embers.rf_tools.colormaps import spectral
+from embers.tile_maps.beam_utils import (chisq_fit_gain,
+                                         healpix_cardinal_slices, map_slices,
+                                         poly_fit, rotate_map)
+from matplotlib import pyplot as plt
+from mpl_toolkits.axes_grid1 import make_axes_locatable
+
+matplotlib.use("Agg")
+_spec, _ = spectral()
+
+parser = argparse.ArgumentParser(
+    description="""
+    Tile Slices paper plot
+    """
+)
+
+parser.add_argument(
+    "--out_dir",
+    metavar="\b",
+    default="../embers_out/paper_plots/tile_slices/",
+    help="Output directory. Default=../embers_out/paper_plots",
+)
+parser.add_argument(
+    "--map_dir",
+    metavar="\b",
+    default="../embers_out/tile_maps/tile_maps/tile_maps_clean/",
+    help="Tile map directory. Default=../embers_out/tile_maps/tile_maps/tile_maps_clean",
+)
+parser.add_argument(
+    "--fee_map",
+    metavar="\b",
+    default="../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz",
+    help="Healpix FEE map of mwa tile. default=../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz",
+)
+parser.add_argument(
+    "--nside", metavar="\b", type=int, default=32, help="Healpix Nside. Default = 32",
+)
+
+args = parser.parse_args()
+
+out_dir = Path(args.out_dir)
+map_dir = Path(args.map_dir)
+fee_map = args.fee_map
+nside = args.nside
+
+# make output dir if it doesn't exist
+out_dir.mkdir(parents=True, exist_ok=True)
+
+tiles = [
+    "S06",
+    "S07",
+    "S08",
+    "S09",
+    "S10",
+    "S12",
+    "S29",
+    "S30",
+    "S31",
+    "S32",
+    "S33",
+    "S34",
+    "S35",
+    "S36",
+]
+
+refs = ["rf1"]
+
+tile_pairs = []
+for r in refs:
+    for t in tiles:
+        tile_pairs.append([t, r])
+
+
+def beam_slices(map_file, fee_map, nside):
+    """Returns pairs of [NS,EW] slices of maps, for each pointing"""
+
+    t_name, r_name, _, _ = Path(map_file).stem.split("_")
+
+    pointings = ["0", "2", "4"]
+
+    maps = []
+
+    # load data from map .npz file
+    tile_map = np.load(map_file, allow_pickle=True)
+    fee_m = np.load(fee_map, allow_pickle=True)
+
+    for p in pointings:
+
+        tile = tile_map[p]
+
+        if "XX" in t_name:
+            fee = fee_m[p][0]
+        else:
+            fee = fee_m[p][1]
+
+        # rotate maps so slices can be taken
+        fee_r = rotate_map(nside, angle=-np.pi / 4, healpix_array=fee)
+        tile_r = rotate_map(nside, angle=-np.pi / 4, healpix_array=tile)
+
+        # slice the tile and fee maps along NS, EW
+        # zenith angle thresh of 70 to determine fit gain factor
+        NS_f, EW_f = healpix_cardinal_slices(nside, fee_r, 70)
+        NS_t, EW_t = map_slices(nside, tile_r, 70)
+
[…]
