### 2020-09-19 5fe196d
Mitch's beam grad idea
