### 2020-09-19 8a0cb16
Modular ploting function. Use to create beam varience arry of plots
