### 2020-09-19 3d1dda5
Array of beam varience plots
