+++ paper_plots/beam_varience.py
@@ -1,6 +1,4 @@
 import argparse
-import concurrent.futures
-from itertools import repeat
 from pathlib import Path
 
 import matplotlib
@@ -10,7 +8,7 @@ from embers.tile_maps.beam_utils import (chisq_fit_gain,
                                          healpix_cardinal_slices, map_slices,
                                          poly_fit, rotate_map)
 from matplotlib import pyplot as plt
-from mpl_toolkits.axes_grid1 import make_axes_locatable
+from matplotlib.lines import Line2D
 
 matplotlib.use("Agg")
 _spec, _ = spectral()
@@ -24,7 +22,7 @@ parser = argparse.ArgumentParser(
 parser.add_argument(
     "--out_dir",
     metavar="\b",
-    default="../embers_out/paper_plots/tile_slices/",
+    default="../embers_out/paper_plots",
     help="Output directory. Default=../embers_out/paper_plots",
 )
 parser.add_argument(
@@ -169,8 +167,7 @@ for tile in tile_pairs:
 
 
 def slice_scatter(
-    fig=None,
-    sub=(None, None, None),
+    ax,
     data=None,
     pol=None,
     xlabel=False,
@@ -186,8 +183,6 @@ def slice_scatter(
         - ax - :func:`~matplotlib.pyplot.subplot` object
 
     """
-
-    ax = fig.add_subplot(sub[0], sub[1], sub[2])
 
     tile_keys = list(data.keys())
     colors = _spec(np.linspace(0.17, 0.9, len(tile_keys)))
@@ -250,63 +245,106 @@ plt.rcParams.update(nice_fonts)
 
 fig1 = plt.figure(figsize=(7.6, 9.0))
 
-plt.subplot(4, 3, 1)
+ax1 = fig1.add_axes([0.00, 0.72, 0.32, 0.21])
 plt.title(r"($i$) NS slice of XX beam [zenith]", loc="left")
 slice_scatter(
-    fig=fig1, sub=[4, 3, 1], data=NS_0, pol="XX",
+    ax1, data=NS_0, pol="XX",
 )
 
-plt.subplot(4, 3, 2)
+ax2 = fig1.add_axes([0.33, 0.72, 0.32, 0.21])
 plt.title(r"($ii$) NS slice of XX beam [2]", loc="left")
-slice_scatter(fig=fig1, sub=[4, 3, 2], data=NS_2, pol="XX", ylabel=False)
+slice_scatter(ax2, data=NS_2, pol="XX", ylabel=False)
 
-plt.subplot(4, 3, 3)
+ax3 = fig1.add_axes([0.66, 0.72, 0.32, 0.21])
 plt.title(r"($iii$) NS slice of XX beam [4]", loc="left")
-slice_scatter(fig=fig1, sub=[4, 3, 3], data=NS_4, pol="XX", ylabel=False)
+slice_scatter(ax3, data=NS_4, pol="XX", ylabel=False)
 
-plt.subplot(4, 3, 4)
+ax4 = fig1.add_axes([0.00, 0.48, 0.32, 0.21])
 plt.title(r"($iv$) EW slice of XX beam [zenith]", loc="left")
 slice_scatter(
-    fig=fig1, sub=[4, 3, 4], data=EW_0, pol="XX",
+    ax4, data=EW_0, pol="XX",
 )
 
-plt.subplot(4, 3, 5)
+ax5 = fig1.add_axes([0.33, 0.48, 0.32, 0.21])
[…]
