### 2020-09-20 3e5ed34
Have computed normal, but makes no physical sense
