### 2020-09-22 87dc5f8
Weird mess trying to rotate the sky
