+++ paper_plots/tilt_fit.py
@@ -1,0 +1,229 @@
+from pathlib import Path
+
+import healpy as hp
+import numpy as np
+from embers.rf_tools.colormaps import jade
+from matplotlib import pyplot as plt
+from healpy.rotator import euler_matrix_new as euler
+
+
+jade, _ = jade()
+
+fee_map = "../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz"
+nside = 32
+
+
+def reproject_map(nside, pixel=None, healpix_array=None):
+    """Reproject beam map to be centered around healpix pixel instead of zenith.
+
+    :param nside: Healpix nside
+    :param pixel: Healpix index of new phase centre
+    :param healpix_array: Input healpix array to be rotated
+
+    :returns:
+        - Reprojected healpix map
+
+    """
+
+    # coordinated of new phase centre
+    theta, phi = hp.pix2ang(nside, pixel)
+
+    # theta phi values of each pixel
+    hp_indices = np.arange(hp.nside2npix(nside))
+    θ, ɸ = hp.pix2ang(nside, hp_indices)
+
+    new_theta = []
+    new_phi = []
+
+    for i in θ - theta:
+        if i >= 0:
+            new_theta.append(i)
+        else:
+            new_theta.append(-i)
+
+    for j in ɸ - phi:
+        if j <= 2 * np.pi:
+            new_phi.append(j)
+        else:
+            new_phi.append(j - 2 * np.pi)
+
+    new_hp_inds = hp.ang2pix(nside, new_theta, ɸ)
+    #  new_hp_inds = hp.ang2pix(nside, θ - theta, ɸ - phi)
+
+    return healpix_array[new_hp_inds]
+
+
+def plot_healpix(
+    data_map=None,
+    rot=(0, 90, 180),
+    fig=None,
+    sub=None,
+    title=None,
+    vmin=-50,
+    vmax=0,
+    cmap=None,
+    cbar=True,
+):
+    """Yeesh do some healpix magic to plot the thing
+
+    :param data_map: Healpix input map to plot
+    :param fig: Figure number to use
+    :param sub: Matplotlib subplot syntax
+    :param title: Plot title
+    :param vmin: Colormap minimum
+    :param vmax: Colormap maximum
+    :param cmap: Matplotlib :class:`~matplotlib.colors.ListedColormap`
+    :param cbar: If True, plot a colorbar
+
+    :returns:
+        - Plot of healpix map
+
+    """
+
+    # Disable cryptic healpy warnings. Can't figure out where they originate
+    import warnings
+
+    warnings.filterwarnings("ignore", category=RuntimeWarning)
+
+    hp.delgraticules()
+    hp.orthview(
+        map=data_map,
+        coord="E",
+        fig=fig,
+        half_sky=True,
+        rot=rot,
+        xsize=1200,
+        title=title,
+        sub=sub,
+        min=vmin,
+        max=vmax,
+        cmap=cmap,
+        notext=True,
+        hold=True,
+        cbar=cbar,
+        return_projected_map=True,
+    )
+
+    hp.graticule(dpar=10, coord="E", color="k", alpha=0.7, dmer=45, lw=0.4, ls=":")
+
+    # Altitude grid
+    hp.projtext(
+        00.0 * (np.pi / 180.0),
+        225.0 * (np.pi / 180),
+        "0",
+        color="k",
+        coord="E",
+        fontsize=6,
+        fontweight="light",
+    )
+    hp.projtext(
+        30.0 * (np.pi / 180.0),
+        225.0 * (np.pi / 180),
+        "30",
+        color="k",
+        coord="E",
[…]
