### 2020-09-22 0ef3444
New directions. Commiting before I nuke old healpy based techniques. Was not accurate enough. Using 2D splines now
