+++ paper_plots/tilt_fit.py
@@ -1,3 +1,4 @@
+import itertools
 from pathlib import Path
 
 import healpy as hp
@@ -5,6 +6,7 @@ import numpy as np
 from embers.rf_tools.colormaps import jade
 from healpy.rotator import euler_matrix_new as euler
 from matplotlib import pyplot as plt
+from scipy import interpolate
 
 jade, _ = jade()
 
@@ -12,7 +14,8 @@ fee_map = "../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz"
 nside = 32
 
 
-def reproject_map(nside, pixel=None, healpix_array=None):
+#  def reproject_map(nside, pixel=None, healpix_array=None):
+def reproject_map(nside, healpix_array=None):
     """Reproject beam map to be centered around healpix pixel instead of zenith.
 
     :param nside: Healpix nside
@@ -25,10 +28,11 @@ def reproject_map(nside, pixel=None, healpix_array=None):
     """
 
     # coordinated of new phase centre
-    theta, phi = hp.pix2ang(nside, pixel)
+    #  theta, phi = hp.pix2ang(nside, pixel)
+    theta, phi = [1, 0]
 
     vec = hp.pix2vec(nside, np.arange(hp.nside2npix(nside)))
-    eu_mat = euler(-phi, -theta, phi)
+    eu_mat = euler(-phi, -theta, phi, deg=True)
     rot_map = hp.rotator.rotateVector(eu_mat, vec)
     new_hp_inds = hp.vec2pix(nside, rot_map[0], rot_map[1], rot_map[2])
 
@@ -192,42 +196,116 @@ def beam_maps(f):
     return maps
 
 
-maps = beam_maps(
+#  maps = beam_maps(
+#  "../embers_out/tile_maps/tile_maps/tile_maps_clean/S06XX_rf0XX_tile_maps.npz"
+#  )[0][0]
+
+#  fee_m = np.load(fee_map, allow_pickle=True)
+#  fee = fee_m["0"][0]
+
+#  plot_healpix(data_map=maps, vmin=-50, vmax=0)
+#  plt.savefig("rot_i.png")
+#  plt.close()
+
+#  residuals = maps - fee
+#  residuals[np.where(fee < -30)] = np.nan
+#  residuals[np.where(maps == np.nan)] = np.nan
+
+#  plot_healpix(data_map=residuals, vmin=-5, vmax=5)
+#  plt.savefig("rot_i_res.png")
+#  plt.close()
+
+#  #  map_rot = reproject_map(nside, 0, healpix_array=maps)
+#  map_rot = reproject_map(nside, healpix_array=maps)
+
+#  residuals = map_rot - fee
+#  residuals[np.where(fee < -30)] = np.nan
+#  residuals[np.where(map_rot == np.nan)] = np.nan
+
+#  plot_healpix(data_map=map_rot, vmin=-50, vmax=0)
+#  plt.savefig("rot_ii.png")
+#  plt.close()
+
+#  plot_healpix(data_map=residuals, vmin=-5, vmax=5)
+#  plt.savefig("rot_ii_res.png")
+#  plt.close()
+
+
+def polyfit2d(x, y, z, order=3):
+    ncols = (order + 1) ** 2
+    G = np.zeros((x.size, ncols))
+    ij = itertools.product(range(order + 1), range(order + 1))
+    for k, (i, j) in enumerate(ij):
+        G[:, k] = x ** i * y ** j
[…]
