### 2020-09-22 2dd824d
Good spline inter and masking
