+++ paper_plots/tilt_fit.py
@@ -1,10 +1,8 @@
-import itertools
 from pathlib import Path
 
 import healpy as hp
 import numpy as np
 from embers.rf_tools.colormaps import jade
-from healpy.rotator import euler_matrix_new as euler
 from matplotlib import pyplot as plt
 from scipy import interpolate
 
@@ -12,152 +10,6 @@ jade, _ = jade()
 
 fee_map = "../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz"
 nside = 32
-
-
-#  def reproject_map(nside, pixel=None, healpix_array=None):
-def reproject_map(nside, healpix_array=None):
-    """Reproject beam map to be centered around healpix pixel instead of zenith.
-
-    :param nside: Healpix nside
-    :param pixel: Healpix index of new phase centre
-    :param healpix_array: Input healpix array to be rotated
-
-    :returns:
-        - Reprojected healpix map
-
-    """
-
-    # coordinated of new phase centre
-    #  theta, phi = hp.pix2ang(nside, pixel)
-    theta, phi = [1, 0]
-
-    vec = hp.pix2vec(nside, np.arange(hp.nside2npix(nside)))
-    eu_mat = euler(-phi, -theta, phi, deg=True)
-    rot_map = hp.rotator.rotateVector(eu_mat, vec)
-    new_hp_inds = hp.vec2pix(nside, rot_map[0], rot_map[1], rot_map[2])
-
-    return healpix_array[new_hp_inds]
-
-
-def plot_healpix(
-    data_map=None,
-    rot=(0, 90, 180),
-    fig=None,
-    sub=None,
-    title=None,
-    vmin=None,
-    vmax=None,
-    cmap=None,
-    cbar=True,
-):
-    """Yeesh do some healpix magic to plot the thing
-
-    :param data_map: Healpix input map to plot
-    :param fig: Figure number to use
-    :param sub: Matplotlib subplot syntax
-    :param title: Plot title
-    :param vmin: Colormap minimum
-    :param vmax: Colormap maximum
-    :param cmap: Matplotlib :class:`~matplotlib.colors.ListedColormap`
-    :param cbar: If True, plot a colorbar
-
-    :returns:
-        - Plot of healpix map
-
-    """
-
-    # Disable cryptic healpy warnings. Can't figure out where they originate
-    import warnings
-
-    warnings.filterwarnings("ignore", category=RuntimeWarning)
-
-    hp.delgraticules()
-    hp.orthview(
-        map=data_map,
-        coord="E",
-        fig=fig,
-        half_sky=True,
-        rot=rot,
-        xsize=1200,
-        title=title,
-        sub=sub,
-        min=vmin,
-        max=vmax,
-        cmap=cmap,
-        notext=True,
-        hold=True,
-        cbar=cbar,
-        return_projected_map=True,
-    )
-
-    hp.graticule(dpar=10, coord="E", color="k", alpha=0.7, dmer=45, lw=0.4, ls=":")
-
-    # Altitude grid
-    hp.projtext(
-        00.0 * (np.pi / 180.0),
-        225.0 * (np.pi / 180),
-        "0",
-        color="k",
-        coord="E",
-        fontsize=6,
-        fontweight="light",
-    )
-    hp.projtext(
-        30.0 * (np.pi / 180.0),
[…]
