### 2020-09-22 1d8b31f
Working on implementing euler rotations
