### 2020-09-23 af59b53
Implemented rbf type of interpolations
