+++ paper_plots/tilt_fit.py
@@ -3,8 +3,10 @@ from pathlib import Path
 import healpy as hp
 import numpy as np
 from embers.rf_tools.colormaps import jade
+from embers.tile_maps.beam_utils import plot_healpix
 from matplotlib import pyplot as plt
 from scipy import interpolate
+from scipy import interpolate as interp
 from scipy.spatial.transform import Rotation as R
 
 jade, _ = jade()
@@ -59,6 +61,7 @@ fee = fee_m["0"][0]
 
 def beam_spline(beam_map):
 
+    # Input data
     pix_ind = np.arange(hp.nside2npix(nside))
     x, y, _ = hp.pix2vec(nside, pix_ind)
     z = beam_map[:6144]
@@ -67,56 +70,87 @@ def beam_spline(beam_map):
     y = y[:6144][~mask]
     z = z[~mask]
 
-    def circular_mask(h, w):
+    #  def circular_mask(h, w):
 
-        center = (int(w / 2), int(h / 2))
-        radius = min(center[0], center[1], w - center[0], h - center[1])
+    #  center = (int(w / 2), int(h / 2))
+    #  radius = min(center[0], center[1], w - center[0], h - center[1])
 
-        Y, X = np.ogrid[:h, :w]
-        dist_from_center = np.sqrt((X - center[0]) ** 2 + (Y - center[1]) ** 2)
+    #  Y, X = np.ogrid[:h, :w]
+    #  dist_from_center = np.sqrt((X - center[0]) ** 2 + (Y - center[1]) ** 2)
 
-        mask = dist_from_center <= radius
-        return mask
+    #  mask = dist_from_center <= radius
+    #  return mask
 
-    xnew_edges, ynew_edges = np.mgrid[-1:1:512j, -1:1:512j]
-    xnew = xnew_edges[:-1, :-1] + np.diff(xnew_edges[:2, 0])[0] / 2.0
-    ynew = ynew_edges[:-1, :-1] + np.diff(ynew_edges[0, :2])[0] / 2.0
+    #  xnew_edges, ynew_edges = np.mgrid[-1:1:512j, -1:1:512j]
+    #  xnew = xnew_edges[:-1, :-1] + np.diff(xnew_edges[:2, 0])[0] / 2.0
+    #  ynew = ynew_edges[:-1, :-1] + np.diff(ynew_edges[0, :2])[0] / 2.0
 
     tck = interpolate.bisplrep(x, y, z, s=5000)
-    znew = interpolate.bisplev(xnew[:, 0], ynew[0, :], tck)
-    mask = circular_mask(znew.shape[0], znew.shape[1])
-    znew[~mask] = np.nan
+    #  znew = interpolate.bisplev(xnew[:, 0], ynew[0, :], tck)
+    #  mask = circular_mask(znew.shape[0], znew.shape[1])
+    #  znew[~mask] = np.nan
 
-    return (xnew, ynew, znew)
+    #  return (xnew, ynew, znew)
+    znew = interpolate.bisplev([1, 0, -1], [0, 1, 0], tck)
+    print(znew)
 
 
-fee_x, fee_y, fee_z = beam_spline(fee)
-beam_x, beam_y, beam_z = beam_spline(b_map)
+#  fee_x, fee_y, fee_z = beam_spline(fee)
+#  beam_x, beam_y, beam_z = beam_spline(b_map)
 
-res_z = beam_z - fee_z
-fee_z[np.isnan(fee_z)] = -80
-mask = np.where(fee_z < -30)
-res_z[mask] = np.nan
+#  res_z = beam_z - fee_z
[…]
