### 2020-09-23 920e7ba
Cleaned up imports and added comments about what this script does
