### 2020-09-23 19d55b2
Exploring Mitche's fft to tile gradient idea
