### 2020-09-23 c239c5c
This now works for nside=256. Just takes a while
