### 2020-09-23 739bef2
Interpolate all beam maps to nside of 256 using radial basis functions
