### 2020-09-23 a02ce7d
Interpolate all beam maps to nside of 256 using radial basis functions
