### 2020-09-23 141a400
Interpolate all beam maps to nside of 256 using radial basis functions
