### 2020-09-23 4204d46
Cleaned up a bunch of irrelevant code. Will go in another script
