+++ paper_plots/interp_beam.py
@@ -2,68 +2,7 @@ from pathlib import Path
 
 import healpy as hp
 import numpy as np
-from healpy.rotator import euler_matrix_new as euler
 from scipy import interpolate as interp
-
-
-def beam_maps(f):
-    """Returns pairs of [AUT,FEE] maps, for each pointing"""
-    f_name, _ = Path(f).name.split(".")
-    t_name, r_name, _, _ = f_name.split("_")
-
-    # pointings = ['0','2','4','41']
-    pointings = ["0", "2", "4"]
-
-    maps = []
-
-    # load data from map .npz file
-    tile_map = np.load(f, allow_pickle=True)
-    fee_m = np.load(fee_map, allow_pickle=True)
-
-    for p in pointings:
-
-        tile = tile_map[p]
-
-        if "XX" in t_name:
-            fee = fee_m[p][0]
-        else:
-            fee = fee_m[p][1]
-
-        # Visualize the tile map and diff map
-        # healpix meadian map
-        tile_med = np.asarray([(np.nanmedian(j) if j != [] else np.nan) for j in tile])
-
-        residuals = tile_med - fee
-        residuals[np.where(fee < -30)] = np.nan
-        residuals[np.where(tile_med == np.nan)] = np.nan
-
-        maps.append([tile_med, residuals])
-
-    return maps
-
-
-def reproject_map(nside, pixel, healpix_array=None):
-    """Reproject beam map to be centered around healpix pixel instead of zenith.
-    :param nside: Healpix nside
-    :param pixel: Healpix index of new phase centre
-    :param healpix_array: Input healpix array to be rotated
-    :returns:
-        - Reprojected healpix map
-    """
-
-    # coordinated of new phase centre
-    #  theta, phi = hp.pix2ang(nside, pixel)
-    # Theta moves primary beam eastward
-    # Phi rotates primary beam anticlockwise. Or, E --> N
-    #  theta, phi = [1, -45]
-    theta, phi = hp.pix2ang(nside, pixel)
-
-    vec = hp.pix2vec(nside, np.arange(hp.nside2npix(nside)))
-    eu_mat = euler(-phi, -theta, phi, deg=False)
-    rot_map = hp.rotator.rotateVector(eu_mat, vec)
-    new_hp_inds = hp.vec2pix(nside, rot_map[0], rot_map[1], rot_map[2])
-
-    return healpix_array[new_hp_inds]
 
 
 def beam_rbf_interp(beam_map, nside, dest_nside):
