### 2020-09-23 67f311c
Added comment to explain purpose of script
