### 2020-09-23 a53df73
Cleaned up a bunch of old functions
