### 2020-09-23 c61e552
New function to characterize pointing offset due to spherical sky projections
