### 2020-09-28 9dda34c
Added comments describing script
