### 2020-09-28 28d7c05
Added comments describing script
