### 2020-09-28 7e8b73d
Nuked the old stuff. Plot bore-sight angles on sky
