+++ paper_plots/tile_tilt.py
@@ -1,3051 +1,52 @@
-"""
-TILE TILT.
-----------
-"""
-
-import argparse
-from pathlib import Path
-
-import healpy as hp
-import matplotlib
+import json
 import numpy as np
-from embers.rf_tools.colormaps import jade
-from matplotlib import pyplot as plt
-from scipy.linalg import lstsq
-
-matplotlib.use("Agg")
-jade, _ = jade()
-
-parser = argparse.ArgumentParser(
-    description="""
-    Tile Maps paper plot
-    """
-)
-
-parser.add_argument(
-    "--out_dir",
-    metavar="\b",
-    default="../embers_out/paper_plots/tile_maps/",
-    help="Output directory. Default=../embers_out/paper_plots/tile_maps",
-)
-parser.add_argument(
-    "--map_dir",
-    metavar="\b",
-    default="../embers_out/tile_maps/tile_maps/tile_maps_clean/",
-    help="Tile map directory. Default=../embers_out/tile_maps/tile_maps/tile_maps_clean",
-)
-parser.add_argument(
-    "--fee_map",
-    metavar="\b",
-    default="../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz",
-    help="Healpix FEE map of mwa tile. default=../embers_out/mwa_utils/mwa_fee/mwa_fee_beam.npz",
-)
-parser.add_argument(
-    "--nside", metavar="\b", type=int, default=32, help="Healpix Nside. Default = 32",
-)
-
-args = parser.parse_args()
-
-out_dir = Path(args.out_dir)
-map_dir = Path(args.map_dir)
-fee_map = args.fee_map
-nside = args.nside
-
-# make output dir if it doesn't exist
-out_dir.mkdir(parents=True, exist_ok=True)
-
-tiles = [
-    "S06",
-    "S07",
-    "S08",
-    "S09",
-    "S10",
-    "S12",
-    "S29",
-    "S30",
-    "S31",
-    "S32",
-    "S33",
-    "S34",
-    "S35",
-    "S36",
-]
-
-refs = ["rf0", "rf1"]
-
-tile_pairs = []
-for r in refs:
-    for t in tiles:
-        tile_pairs.append([t, r])
+import healpy as hp
+import matplotlib.pyplot as plt
 
 
-def beam_maps(f):
-    """Returns pairs of [AUT,FEE] maps, for each pointing"""
-    f_name, _ = Path(f).name.split(".")
-    t_name, r_name, _, _ = f_name.split("_")
+nside = 256
+data = "./interp_maps/pix_resi.json"
 
-    pointings = ["0", "2", "4"]
+delta_az = 83.1912
 
-    PB_0 = [
-        0,
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10,
-        11,
-        12,
-        13,
-        14,
-        15,
-        16,
-        17,
-        18,
-        19,
-        20,
-        21,
-        22,
-        23,
-        24,
-        25,
-        26,
-        27,
-        28,
-        29,
-        30,
-        31,
-        32,
-        33,
-        34,
-        35,
-        36,
-        37,
-        38,
-        39,
-        40,
-        41,
-        42,
-        43,
-        44,
-        45,
-        46,
-        47,
-        48,
-        49,
-        50,
-        51,
-        52,
-        53,
-        54,
-        55,
-        56,
-        57,
-        58,
-        59,
-        60,
-        61,
-        62,
-        63,
-        64,
-        65,
-        66,
-        67,
-        68,
-        69,
-        70,
-        71,
-        72,
-        73,
-        74,
-        75,
-        76,
-        77,
-        78,
-        79,
[…]
