+++ paper_plots/tile_tilt/tile_tilt.py
@@ -3,10 +3,10 @@ import json
 import healpy as hp
 import numpy as np
 from embers.rf_tools.colormaps import jade, spectral
-from embers.tile_maps.beam_utils import plot_healpix
+#  from embers.tile_maps.beam_utils import plot_healpix
 from matplotlib import pyplot as plt
 
-from tilt_fit import reproject_map
+#  from tilt_fit import reproject_map
 
 _spec, _ = spectral()
 _jade, _ = jade()
@@ -17,6 +17,9 @@ out_dir = "./interp_maps/beam_offsets"
 m_256 = "./interp_maps/0"
 m_32 = "../embers_out/tile_maps/tile_maps/tile_maps_clean"
 
+ref_p = 243.17
+ref_t = 0.476
+
 
 with open(data) as f:
     tilt_resi = json.load(f)
@@ -26,7 +29,6 @@ phase_cen = {}
 
 for tile in tilt_resi.keys():
     if "rf1" in tile and "_0" in tile:
-        print(tile)
         min_pix = np.argmax(tilt_resi[tile])
         phase_cen[f"{tile}"] = np.array(hp.pix2ang(nside, min_pix))
 
@@ -52,75 +54,77 @@ plt.style.use("seaborn")
 figure = plt.figure(figsize=(6, 6))
 ax = figure.add_subplot(111, polar=True)
 
-ax.set_ylim(0, 1.8)
-ax.set_rgrids([0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.2, 1.4, 1.6], angle=90)
+#  ax.set_ylim(0, 1.5)
+#  ax.set_rgrids([0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.2, 1.4], angle=90)
 ax.set_theta_zero_location("N")
 ax.set_theta_direction(-1)
 
 for i, c in enumerate(colors):
+    #  plt.scatter(
+    #  (phi_YY[i]),
+    #  np.degrees(theta_YY[i]),
+    #  marker="P",
+    #  color=c,
+    #  edgecolor="black",
+    #  linewidth=0.7,
+    #  s=77,
+    #  )
+    #  plt.scatter(
+    #  (phi_XX[i]),
+    #  np.degrees(theta_XX[i]),
+    #  marker="X",
+    #  color=c,
+    #  edgecolor="black",
+    #  linewidth=0.7,
+    #  s=77,
+    #  )
     plt.scatter(
-        phi_YY[i],
-        np.degrees(theta_YY[i]),
-        marker="P",
-        color=c,
-        edgecolor="black",
-        linewidth=0.7,
-        s=77,
-    )
-    plt.scatter(
-        phi_XX[i],
-        np.degrees(theta_XX[i]),
-        marker="X",
-        color=c,
-        edgecolor="black",
-        linewidth=0.7,
-        s=77,
-    )
-    plt.scatter(
-        phi[i],
+        (phi[i]),
         np.degrees(theta[i]),
-        marker="o",
+        marker="s",
         color=c,
         edgecolor="black",
         linewidth=0.7,
-        s=77,
+        s=84,
+        alpha=0.9,
     )
 #  plt.legend()
+ax.set_xticklabels(["N", "", "E", "", "S", "", "W", ""])
 plt.tight_layout()
 plt.savefig("./interp_maps/beam_offsets/polar_offsets.png")
[…]
