### 2020-09-30 f41b6b4
Cleaner tile tilt plot
