### 2020-09-30 7c61c38
Plot tilt of southern hex and find gradient
