### 2020-09-30 c3f5997
Plot ready for the paper
