+++ paper_plots/tile_tilt/tile_tilt.py
@@ -37,6 +37,7 @@ tiles = [
     "S34",
     "S35",
     "S36",
+    "Soil",
 ]
 
 
@@ -66,6 +67,8 @@ for i in list(phase_cen.keys()):
 theta = np.mean([theta_XX, theta_YY], axis=0)
 phi = np.mean([phi_XX, phi_YY], axis=0)
 
+hex_phi = np.arctan(72.758 / 84.01) + np.pi / 2
+hex_the = np.degrees(np.arctan(1.016 / np.sqrt(72.758 ** 2 + 84.01 ** 2)))
 
 colors = _spec(np.linspace(0.17, 0.9, len(theta_XX)))
 
@@ -85,6 +88,19 @@ ti = [
     for c in colors
 ]
 
+ti.append(
+    Line2D(
+        [0],
+        [0],
+        color="black",
+        linestyle="None",
+        marker="X",
+        markerfacecolor="black",
+        markersize=7,
+        markeredgecolor="black",
+        markeredgewidth=0.4,
+    )
+)
 
 plt.style.use("seaborn")
 
@@ -139,16 +155,20 @@ for i, c in enumerate(colors):
         s=84,
         alpha=0.9,
     )
+
+plt.scatter(hex_phi, hex_the, marker="X", color="black", s=121, alpha=0.9)
 ax.set_xticklabels(["N", "", "E", "", "S", "", "W", ""])
 
 ax2 = figure.add_axes([0.08, 0.02, 0.84, 0.12])
 ax2.axis("off")
 ax2.legend(ti, tiles, mode="expand", ncol=5, frameon=True)
-plt.savefig("./interp_maps/beam_offsets/polar_offsets.png")
+plt.savefig("./interp_maps/beam_offsets/polar_offsets.pdf", bbox_inches="tight")
 plt.close()
 
 
-#  m_06_32 = "../../embers_out/tile_maps/tile_maps/tile_maps_clean/S08XX_rf1XX_tile_maps.npz"
+#  m_06_32 = (
+#  "../../embers_out/tile_maps/tile_maps/tile_maps_clean/S08XX_rf1XX_tile_maps.npz"
+#  )
 #  m_06_32 = np.load(m_06_32, allow_pickle=True)["0"]
 #  m_06_32 = np.asarray([(np.nanmedian(j) if j != [] else np.nan) for j in m_06_32])
 
@@ -180,4 +200,4 @@ plt.close()
 #  plot_healpix(data_map=res_256_rot, vmin=-4, vmax=4, cmap="RdYlGn", cbar=False)
 
 #  plt.tight_layout()
-#  plt.savefig(f"{out_dir}/S08XX_rf1_0.png")
+#  plt.savefig(f"./interp_maps/beam_offsets/tilt_demo.png")
