+++ paper_plots/tile_tilt/hex_tilt.py
@@ -97,5 +97,5 @@ divider = make_axes_locatable(ax)
 cax = divider.append_axes("right", size="5%", pad=0.1)
 fig.colorbar(im, cax=cax)
 plt.tight_layout()
-plt.show()
+#  plt.show()
 plt.savefig("./interp_maps/beam_offsets/hex_tilt.png")
