### 2020-10-07 1b8f4a9
Only use zenith pointing
