### 2020-10-07 facf966
Investigate rotation of references using feko sims
