### 2020-10-07 c7062bd
Error bars, mask primary beam
