+++ paper_plots/ref_rot.py
@@ -1,25 +1,85 @@
 import healpy as hp
 import numpy as np
 from embers.rf_tools.colormaps import jade, spectral
-from embers.tile_maps.beam_utils import plot_healpix, rotate_map
+from embers.tile_maps.beam_utils import (healpix_cardinal_slices, plot_healpix,
+                                         rotate_map)
+from healpy.rotator import euler_matrix_new as euler
 from matplotlib import pyplot as plt
 
-spec, _ = spectral()
-jade, _ = jade()
-nside = 32
 
-ref_model = "../embers_out/tile_maps/ref_models/ref_dipole_models.npz"
+def reproject_map(nside, phi, healpix_array=None):
+    """Reproject beam map to be centered around healpix pixel instead of zenith.
 
-# Load reference FEE model
-# Rotate the fee models by -pi/2 to move model from spherical (E=0) to Alt/Az (N=0)
-ref_fee_model = np.load(ref_model, allow_pickle=True)
+    :param nside: Healpix nside
+    :param phi: Angle to rotate beam by, from E --> N (anticlockwise)
+    :param healpix_array: Input healpix array to be rotated
+    :returns:
+        - Reprojected healpix map
+    """
 
-ref_xx = rotated_fee = rotate_map(
-    nside, angle=-(1 * np.pi) / 2.0, healpix_array=ref_fee_model["XX"]
-)
-ref_yy = rotated_fee = rotate_map(
-    nside, angle=-(1 * np.pi) / 2.0, healpix_array=ref_fee_model["YY"]
-)
+    vec = hp.pix2vec(nside, np.arange(hp.nside2npix(nside)))
+    eu_mat = euler(-phi, 0, 0, deg=True)
+    rot_map = hp.rotator.rotateVector(eu_mat, vec)
+    new_hp_inds = hp.vec2pix(nside, rot_map[0], rot_map[1], rot_map[2])
 
-plot_healpix(data_map=ref_xx, cmap=spec, vmin=-20, vmax=0)
-plt.savefig("test_ref.png")
+    return healpix_array[new_hp_inds]
+
+
+if __name__ == "__main__":
+
+    spec, _ = spectral()
+    jade, _ = jade()
+    nside = 512
+
+    ref_model = "ref_rot/ref_dipole_models.npz"
+
+    # Load reference FEE model
+    # Rotate the fee models by -pi/2 to move model from spherical (E=0) to Alt/Az (N=0)
+    ref_fee_model = np.load(ref_model, allow_pickle=True)
+
+    ref_xx = rotated_fee = rotate_map(
+        nside, angle=-(1 * np.pi) / 2.0, healpix_array=ref_fee_model["XX"]
+    )
+    ref_yy = rotated_fee = rotate_map(
+        nside, angle=-(1 * np.pi) / 2.0, healpix_array=ref_fee_model["YY"]
+    )
+
+    ref_xx_rot = reproject_map(nside, 2, healpix_array=ref_xx)
+
+    res = ref_xx - ref_xx_rot
+
+    NS, EW = healpix_cardinal_slices(nside, res, 85)
+
+    # Plotting stuff
+    plt.style.use("seaborn")
+
+    nice_fonts = {
+        "font.family": "sans-serif",
+        "axes.labelsize": 8,
+        "axes.titlesize": 9,
+        "font.size": 8,
+        "legend.fontsize": 8,
+        "xtick.labelsize": 8,
+        "ytick.labelsize": 8,
+    }
+
+    plt.rcParams.update(nice_fonts)
+
+    plt.figure(figsize=(6, 5))
+
+    plt.plot(NS[1], NS[0], label="NS_XX")
[…]
