### 2020-10-10 37f6f45
Interpolate beam maps
