+++ paper_plots/ref_rot.py
@@ -2,9 +2,10 @@ import healpy as hp
 import numpy as np
 from embers.rf_tools.colormaps import jade, spectral
 from embers.tile_maps.beam_utils import (healpix_cardinal_slices, plot_healpix,
-                                         rotate_map, poly_fit)
+                                         poly_fit, rotate_map)
 from healpy.rotator import euler_matrix_new as euler
 from matplotlib import pyplot as plt
+from matplotlib.lines import Line2D
 
 
 def reproject_map(nside, phi, healpix_array=None):
@@ -44,12 +45,6 @@ if __name__ == "__main__":
         nside, angle=-(1 * np.pi) / 2.0, healpix_array=ref_fee_model["YY"]
     )
 
-    ref_xx_rot = reproject_map(nside, 2, healpix_array=ref_xx)
-
-    res = ref_xx - ref_xx_rot
-
-    NS, EW = healpix_cardinal_slices(nside, res, 85)
-
     # Plotting stuff
     plt.style.use("seaborn")
 
@@ -58,29 +53,121 @@ if __name__ == "__main__":
         "axes.labelsize": 8,
         "axes.titlesize": 9,
         "font.size": 8,
-        "legend.fontsize": 8,
+        "legend.fontsize": 5,
         "xtick.labelsize": 8,
         "ytick.labelsize": 8,
     }
 
     plt.rcParams.update(nice_fonts)
 
-    plt.figure(figsize=(6, 5))
+    fig = plt.figure(figsize=(3.6, 2.4))
 
-    plt.plot(NS[1], NS[0], label="NS_XX")
-    plt.plot(EW[1], EW[0], label="EW_XX")
-    plt.plot(NS[1][::36], poly_fit(NS[1][::36], NS[0][::36], NS[0][::36], 9), label="NS_XX")
-    plt.plot(EW[1][::36], poly_fit(EW[1][::36], EW[0][::36], EW[0][::36], 9), label="EW_XX")
+    #  rots = [-30, -10, -8, -6, -4, -2, 0, 2, 4, 6, 8, 10, 30]
+    rots = [12, 9, 6, 3, 0]
+    #  colors = spec(np.linspace(0.14, 0.63, len(rots)))
+    colors = spec([0.14, 0.21, 0.35, 0.63, 0.70])
+    legs = [
+        Line2D([0], [0], linestyle="None", marker="$NS$", markerfacecolor="k", markersize=9),
+        Line2D([0], [0], linestyle="None", marker="$EW$", markerfacecolor="k", markersize=9.6),
+        Line2D([0], [0], color=colors[0], linewidth=2,),
+        Line2D(
+            [0],
+            [0],
+            color=colors[0],
+            linewidth=2,
+            linestyle="dotted",
+            dash_capstyle="round",
+        ),
+        Line2D([0], [0], color=colors[1], linewidth=2,),
+        Line2D(
+            [0],
+            [0],
+            color=colors[1],
+            linewidth=2,
+            linestyle="dotted",
+            dash_capstyle="round",
+        ),
+        Line2D([0], [0], color=colors[2], linewidth=2,),
+        Line2D(
+            [0],
+            [0],
+            color=colors[2],
+            linewidth=2,
+            linestyle="dotted",
+            dash_capstyle="round",
+        ),
+        Line2D([0], [0], color=colors[3], linewidth=2,),
+        Line2D(
+            [0],
+            [0],
+            color=colors[3],
+            linewidth=2,
+            linestyle="dotted",
+            dash_capstyle="round",
+        ),
+        Line2D([0], [0], color=colors[4], linewidth=2,),
+        Line2D(
+            [0],
+            [0],
[…]
