### 2020-10-10 f50ddba
Black outline to legend markers. Increased visibility + match with plot
