### 2020-10-10 c801a9c
Save reference residuals to .npy file to be used for errorbars in beam slice plots
