### 2020-10-10 b1382f7
Change default max za to 90 == horizon
