### 2020-10-10 269c654
Flipped residuals to match convention and moved position of legend box
