### 2020-10-12 9fc3261
Better colormap
