### 2020-10-13 fa3918a
Better colormap
