### 2020-10-13 fab6e34
Combined two rot_res plots. It's a beauty
