+++ paper_plots/cardinal_errors.py
@@ -7,17 +7,21 @@
 import argparse
 from pathlib import Path
 
+import healpy as hp
 import matplotlib
 import numpy as np
 from astropy.stats import mad_std as mad
-from embers.rf_tools.colormaps import spectral
+from embers.rf_tools.colormaps import jade, spectral
 from embers.tile_maps.beam_utils import (chisq_fit_gain,
                                          healpix_cardinal_slices, map_slices,
                                          poly_fit, rotate_map)
+from healpy.rotator import euler_matrix_new as euler
 from matplotlib import pyplot as plt
+from matplotlib.lines import Line2D
 
 matplotlib.use("Agg")
 _spec, _ = spectral()
+
 
 parser = argparse.ArgumentParser(
     description="""
@@ -121,8 +125,8 @@ def beam_slices(map_file, fee_map, nside):
 
         # slice the tile and fee maps along NS, EW.
         # the above gain factor is applied to full beam slices
-        NS_fee, EW_fee = healpix_cardinal_slices(nside, fee_r, 90)
-        NS_tile, EW_tile = map_slices(nside, tile_r, 90)
+        NS_fee, EW_fee = healpix_cardinal_slices(nside, fee_r, 85)
+        NS_tile, EW_tile = map_slices(nside, tile_r, 85)
 
         # Scale the data so that it best fits the beam slice
         NS_tile_med = NS_tile[0] - gain_NS[0]
@@ -144,6 +148,24 @@ def beam_slices(map_file, fee_map, nside):
         )
 
     return maps
+
+
+def reproject_map(nside, phi, healpix_array=None):
+    """Reproject beam map to be centered around healpix pixel instead of zenith.
+
+    :param nside: Healpix nside
+    :param phi: Angle to rotate beam by, from E --> N (anticlockwise)
+    :param healpix_array: Input healpix array to be rotated
+    :returns:
+        - Reprojected healpix map
+    """
+
+    vec = hp.pix2vec(nside, np.arange(hp.nside2npix(nside)))
+    eu_mat = euler(-phi, 0, 0, deg=True)
+    rot_map = hp.rotator.rotateVector(eu_mat, vec)
+    new_hp_inds = hp.vec2pix(nside, rot_map[0], rot_map[1], rot_map[2])
+
+    return healpix_array[new_hp_inds]
 
 
 def slice_residual():
@@ -216,21 +238,21 @@ def slice_residual():
     EW_YY_fit = poly_fit(za, EW_YY_res, EW_YY_res, 2)
 
     # Plotting stuff
-    plt.style.use("seaborn")
+    #  plt.style.use("seaborn")
 
-    nice_fonts = {
-        "font.family": "sans-serif",
-        "axes.labelsize": 8,
-        "axes.titlesize": 9,
-        "font.size": 8,
-        "legend.fontsize": 6,
-        "xtick.labelsize": 8,
-        "ytick.labelsize": 8,
-    }
+    #  nice_fonts = {
+    #  "font.family": "sans-serif",
+    #  "axes.labelsize": 8,
+    #  "axes.titlesize": 9,
+    #  "font.size": 8,
+    #  "legend.fontsize": 6,
+    #  "xtick.labelsize": 8,
+    #  "ytick.labelsize": 8,
+    #  }
 
-    plt.rcParams.update(nice_fonts)
+    #  plt.rcParams.update(nice_fonts)
 
-    plt.figure(figsize=(3.6, 2.4))
+    #  plt.figure(figsize=(3.6, 2.4))
 
     #  colors = _spec(np.linspace(0.17, 0.9, 4))
     colors = _spec([0.14, 0.77, 0.66, 0.35])
[…]
