### 2020-10-14 dc6237a
Better colormap
