### 2020-10-14 26345d5
Better colormap
