### 2020-10-15 #2629 @mbobra
👋 @amanchokshi I am going through the checklist and your package looks really nice. It has been an easy review so far. I have a couple comments:

**Functionality**
* I am trying to follow the [Embers by Example](https://embers.readthedocs.io/en/latest/embersbyexample.html) guide and am stuck on the second code block. Is test_tool installed with embers? When I try to run the command `python -m test_tool`, I get the error: `No module named test_tool`.

**Documentation**

* I am a bit confused about the statement of need in the documentation, since it reads slightly differently than the JOSS paper (I am not an expert in radio astronomy).

From the online documentation:
> EMBERS is a python package designed to enable polarised measurements of radio telescope antenna beam-patterns using satellites.

From the JOSS paper:
> EMBERS is a python package which provides a modular framework for radio telescopes and interferometric arrays such as the MWA, HERA, and the upcoming SKA to accurately measure the all sky beam responses of their antennas using weather and communication satellites.

Could you perhaps clarify and expand the statement of need in the documentation? Also, the documentation doesn't mention the target audience for this software (e.g. astronomers, systems engineers?). Could you please include this?

**Software Paper**
* The statement of need includes this sentence: "Chokshi et al (in prep) represents the first complete large scale implementation of a satellite based beam measurement system." Could you please write one or two sentences about what has been done in the past, why this is an improvement, and how this software compares to other commonly-used packages?
* Could you please write a sentence to the explain the gray (unobserved) areas in the caption of Figure 1?
* We prefer citations with a DOI. The paper includes a couple references to Chokshi et al., in prep, which isn't yet peer-reviewed and doesn't yet have a DOI. Would it be possible to replace this with another reference or update it with a DOI?
