### 2020-10-16 605cfdf
Working on Monica's comments. Added lots of description to the caption of fig 1
