### 2020-10-16 4ba35d9
Changed description to match joss paper intro
