+++ docs/index.rst
@@ -10,7 +10,7 @@
 
    *Experimental Measurement of BEam Responses with Satellites*
 
-EMBERS is a python package designed to enable polarised measurements of radio telescope antenna beam-patterns using satellites. The methodology of the experimental setup can be found in the following paper [Coming soon].
+EMBERS is a python package which provides a modular framework for radio telescopes and interferometric arrays such as the MWA, HERA, and the upcoming SKA-Low to accurately measure the all sky polarised beam responses of their antennas using weather and communication satellites. This tool was designed to enable astronomers and system engineers, all over the world, to characterize the in-situ antenna beam patterns of large arrays with ease. The methodology of the experimental setup can be found in the following paper [Coming soon].
 
     - Written in pure Python and installs without any compilation
     - Supports Python 3.6-3.9
