### 2020-10-16 3351913
Made instructions of usage of cli tools clearer
