+++ docs/embersbyexample.rst
@@ -17,14 +17,16 @@ Setup
 
 The above creates an :samp:`EMBERS` directory, within which we can run our example code. Outputs will by default, be saved to :samp:`~/EMBERS/embers_out`.
 
-A set of cli-tools enable easy interactions with the *EMBERS* packages, each of which can be executed with either of the following methods
+A set of cli-tools enable easy interactions with the *EMBERS* packages. Each of the available tools can be executed with either of the following methods.
+The sample below demonstartes the usage of an example tool called :samp:`test_tool`. 
+
 
 .. code-block:: console
 
     $ python -m test_tool
     $ test_tool
 
-These tools come built with help functions, which can be accessed at any time with the :samp:`--help` flag
+Each tools come built with help functions, which can be accessed at any time with the :samp:`--help` flag
 
 .. code-block:: console
 
