### 2020-10-16 d6ce087
Made instructions of usage of cli tools clearer
