### 2020-10-16 #2629 @amanchokshi
@mbobra Thanks Monica. I've addressed most of your comments.

#### Functionality
In Embers by Example, the second and third code blocks are meant for people unfamiliar with python. `test_tool` is an example package which doesn't exist, but is used to demonstrate that the tools can be run using either `python -m test_tool` or simply `test_tool`. The following code block show that help can be accessed for each tool using the `--help` flag. I've made the preceding text clearer by saying that this is an example tool.

#### Documentation
I've updated the statement of need in the online docs to match the JOSS paper and have included the target audience.

#### Paper
Added that all previous experiments were validations of the methodology, and this was the first large scale test. There was no previously available software, suitable for the large data volume. I've also update the caption of Fig. 1 with much more detail on what all the plots mean, including the grey regions.

Unfortunately, Chokshi et al., in prep is under collaboration review right now, and will probably take a couple of months before it's published. The other paper which is relevant is Line et al., which I've cited. It's about one of the previous validation experiments and is not as relevant as it deals with much smaller scale and does not use the EMBERS package.
