### 2020-10-16 #2629 @mbobra
Thanks for your updates, I really appreciate it. I'm working through the examples and am still a bit confused.

One example the Embers by Example docs says:
> To get a quick preview of the raw RF data, we create waterfall plots. The following code creates a waterfall plot of sample data provided with EMBERS
```python
$ waterfall_single
--------------------------------------------------
No input data provided, using packaged sample data
>>> waterfall_single --help, for more options
--------------------------------------------------
Waterfall plot saved to ./embers_out/rf_tools/S06XX_2019-10-10-02:30.png
```

But this doesn't work exactly as written for me; I have to do a
```python
from embers.rf_tools import rf_data
rf_data.single_waterfall
```
But then I'm stuck, because I don't see any example rf_tiles to use with `rf_data.single_waterfall()`.

Would it be possible to create a couple examples that a user can execute from start to finish? You have a really nice example gallery, so by no means am I asking to re-do all of it, but I think it is a bit confusing on how exactly to get started. Some sample data may be a nice way to quickly show users how they can use the package.
