### 2020-10-19 cf0982c
Added src to MANIFEST.in
