+++ build/lib/embers/tile_maps/null_test.py
@@ -1,0 +1,499 @@
+"""
+Null Tests
+----------
+
+A set of tools to perform null tests on reference rf data and reference beam models
+
+"""
+
+from pathlib import Path
+
+import healpy as hp
+import matplotlib
+import numpy as np
+from embers.tile_maps.beam_utils import (chisq_fit_gain,
+                                         healpix_cardinal_slices, map_slices,
+                                         plt_slice, poly_fit, rotate_map)
+from matplotlib import pyplot as plt
+
+matplotlib.use("Agg")
+
+
+def good_ref_maps(nside, map_dir, tile_pair):
+    """Collates reference data from 18 good satellites into a good_ref_map
+
+    :param nside: Healpix nside
+    :param map_dir: Path to directory with tile_maps_raw, created by :func:`~embers.tile_maps.tile_maps.project_tile_healpix`
+    :param tile_pair: List of a pair of mwa tile and reference names. Ex: ["S35XX", "rf0XX"]
+
+    :returns:
+        - good_ref_map healpix map, with each pixel of the healpix map containing an array of values from all satellite passes within the pixel
+
+    """
+
+    pointings = ["0", "2", "4", "41"]
+
+    # load data from map .npz file
+    f = Path(f"{map_dir}/{tile_pair[0]}_{tile_pair[1]}_sat_maps.npz")
+    tile_data = np.load(f, allow_pickle=True)
+    tile_data = {key: tile_data[key].item() for key in tile_data}
+    ref_map = tile_data["ref_map"]
+
+    # Good sats from which to make plots
+    good_sats = [
+        25338,
+        25984,
+        25985,
+        28654,
+        40086,
+        40087,
+        40091,
+        41179,
+        41180,
+        41182,
+        41183,
+        41184,
+        41185,
+        41187,
+        41188,
+        41189,
+        44387,
+    ]
+
+    # Empty good ref map
+    good_ref_map = [[] for pixel in range(hp.nside2npix(nside))]
+
+    for p in pointings:
+
+        # append to good ref map from all good sat data
+        for sat in good_sats:
+            for pix in range(hp.nside2npix(nside)):
+                good_ref_map[pix].extend(ref_map[p][sat][pix])
+
+    return good_ref_map
+
+
+def plt_null_test(
+    fig=None,
+    sub=(None, None, None),
+    zen_angle=None,
+    del_pow=None,
+    del_err=None,
+    del_beam=None,
+    del_fit=None,
+    null_label=None,
+    beam_label=None,
+    fit_label=None,
+    ylabel=True,
+    title=None,
+):
+
+    """Plot null test between two corresponding slices of reference beam maps
+
+    :param fig: Figure number
+    :param sub: Subplot position, tuple of matplotlib indices. Ex: (1, 1, 1)
+    :param zen_angle: Array of zenith angles
[…]
