+++ build/lib/embers/sat_utils/sat_channels.py
@@ -1,0 +1,760 @@
+"""
+Satellite Channels
+------------------
+
+A set of tools to determine the transmission channel of various satellites in 30 minute
+observation by using rf data in conjunctin with chronological satellite ephemeris data.
+
+"""
+
+import concurrent.futures
+import json
+import re
+from itertools import repeat
+from pathlib import Path
+
+import matplotlib as mpl
+import numpy as np
+from embers.rf_tools.colormaps import spectral
+from embers.rf_tools.rf_data import time_tree
+from matplotlib import pylab as pl
+from matplotlib import pyplot as plt
+from scipy.stats import median_absolute_deviation as mad
+
+mpl.use("Agg")
+
+
+def read_aligned(ali_file=None):
+    """Read aligned data from :func:`~embers.rf_tools.align_data.save_aligned` :samp:`npz` file
+
+    :param ali_file: path to a :func:`~embers.rf_tools.align_data.save_aligned` :samp:`npz` file :class:`~str`
+
+    :returns:
+        A :class:`~tuple` (power, times)
+
+        - ref_pow: a smoothed reference power array :class:`~numpy.ndarry`
+        - tile_pow: a smoothed tile power array :class:`~numpy.ndarry`
+        - times: a regular time array :class:`~numpy.ndarry`
+
+    """
+
+    paired_data = np.load(ali_file, allow_pickle=True)
+
+    ref_pow = paired_data["ref_ali"]
+    tile_pow = paired_data["tile_ali"]
+    times = paired_data["time_array"]
+
+    return (ref_pow, tile_pow, times)
+
+
+def noise_floor(sat_thresh, noi_thresh, power):
+    """Computes the noise floor of a rf power array
+
+    Exclude channels with signal above :samp:`sat_thresh` multiplied by :samp:`standard deviation` of power array.
+    The Median Absolute Deviation :samp:`MAD` is used to quantify the noise level of the remaining
+    channels. The noise floor :samp:`noi_thresh` is defined to be the :samp:`median` of noisy data + :samp:`noi_thresh` multiplied by the
+    :samp:`MAD` of noisy data.
+
+    :param sat_thresh: An integer multiple of standard deviation of rf power array, used to exclude channels with potential satellites. :class:`~int`
+    :param noi_thresh: An integer multiple of the noisy data MAD, used to compute a noise floor. :class:`~int`
+    :param power: Rf power array :class:`~numpy.ndarry`
+
+    :returns:
+        noise_threshold: The power level of the noise floor in dBm :class:`~int`
+
+    """
+
+    # compute the standard deviation of data, and use it to identify occupied channels
+    σ = np.std(power)
+
+    # Any channel with a max power >= σ has a satellite
+    sat_cut = sat_thresh * σ
+    chans_pow_max = np.amax(power, axis=0)
+
+    # Exclude the channels with sats, to only have noise data
+    noise_chans = np.where(chans_pow_max < sat_cut)[0]
+    noise_data = power[:, noise_chans]
+
+    # noise median, noise mad, noise threshold = μ + 3*σ
+    μ_noise = np.median(noise_data)
+    σ_noise = mad(noise_data, axis=None)
+    noise_threshold = μ_noise + noi_thresh * σ_noise
+
+    return noise_threshold
+
+
+def time_filter(s_rise, s_set, times):
+    """Determine indices of time array when a satellite is above the horizon.
+
+    Isolate the portion of a rf power array where a satellite is above the
+    horizon using the rise and set times of a satellite's ephemeris. This
+    function returns a pair of indices which can be used to slice the rf
+    power and times arrays to precisely only include the satellite.
+
+    :param s_rise: satellite rise time from ephemeris :class:`~float`
[…]
