+++ build/lib/embers/mwa_utils/__init__.py
@@ -1,0 +1,6 @@
+"""
+:mod:`embers.mwa_utils` is used to download and metadata of the `MWA Telescope <http://www.mwatelescope.org/>`_ and compute FEE beam models.
+
+It contains :mod:`~embers.mwa_utils.mwa_pointings`, :mod:`~embers.mwa_utils.mwa_dipoles` , :mod:`~embers.mwa_utils.mwa_fee`
+
+"""

[…]
