+++ build/lib/embers/rf_tools/colormaps.py
@@ -1,391 +1,0 @@
-"""
-Colormaps
----------
-A set of custom colormaps used by embers
-
-"""
-
-import numpy as np
-from matplotlib import pyplot as plt
-from matplotlib.colors import ListedColormap
-from mpl_toolkits.axes_grid1 import make_axes_locatable
-from scipy.interpolate import interp1d
-
-
-def spectral():
-    """Beautiful non-linear spectral colormap
-
-    :func:`~embers.rf_tools.colormaps.spectral` is not perceptually uniform and is
-    only used to easily preview raw data with high contrast
-
-    :returns:
-        - spectral, spectral_r - ember colormap :class:`~matplotlib.colors.ListedColormap`.
-
-
-    """
-
-    A = np.array([219 / 256, 55 / 256, 82 / 256, 1])
-    B = np.array([226 / 256, 57 / 256, 79 / 256, 1])
-    C = np.array([230 / 256, 60 / 256, 76 / 256, 1])
-    D = np.array([244 / 256, 109 / 256, 67 / 256, 1])
-    E = np.array([253 / 256, 174 / 256, 97 / 256, 1])
-    F = np.array([254 / 256, 224 / 256, 139 / 256, 1])
-    G = np.array([230 / 256, 245 / 256, 152 / 256, 1])
-    H = np.array([171 / 256, 221 / 256, 164 / 256, 1])
-    I = np.array([102 / 256, 194 / 256, 165 / 256, 1])
-    J = np.array([50 / 256, 136 / 256, 189 / 256, 1])
-    K = np.array([94 / 256, 79 / 256, 162 / 256, 1])
-    L = np.array([46 / 256, 67 / 256, 92 / 256, 1])
-    M = np.array([38 / 256, 42 / 256, 77 / 256, 1])
-    N = np.array([37 / 256, 43 / 256, 61 / 256, 1])
-
-    ncmap = [N, M, L, K, J, I, H, G, F, E, D, C, B, A]
-    c_array = []
-
-    for i in range(len(ncmap)):
-        if i != len(ncmap) - 1:
-            linfit = interp1d([1, 256], np.vstack([ncmap[i], ncmap[i + 1]]), axis=0)
-            for j in range(255):
-                c_array.append(linfit(j + 1))
-
-    # newcmp = ListedColormap(c_array)
-    spec_cmap = ListedColormap(c_array, name="spec", N=len(c_array))
-    spec_cmap_r = spec_cmap.reversed()
-
-    return [spec_cmap, spec_cmap_r]
-
-
-def jade():
-    """Beautiful perceptually uniform jade green colormap.
-
-    :returns:
-        - jade, jade_r - ember colormap :class:`~matplotlib.colors.ListedColormap`
-
-    """
-
-    colorlist = [
-        [0.08048525330056805, 0.05201415138494773, 0.18107886855112176],
-        [0.08282174108497872, 0.05548264112853812, 0.18701136971548327],
-        [0.08510083379013808, 0.05893183335200911, 0.19295110050577835],
-        [0.08732149679165027, 0.06236391678238993, 0.1988999668728864],
-        [0.08948351372779442, 0.06578120659935585, 0.2048568371507506],
-        [0.09158630556914435, 0.06918580148954842, 0.21082119130300758],
-        [0.09362842795942464, 0.07257954032276293, 0.21679429113373555],
-        [0.09560935498577233, 0.07596440682459804, 0.22277440676148685],
-        [0.09752794289012831, 0.07934220307926157, 0.22876097274209006],
-        [0.09938284674215636, 0.08271467009120442, 0.23475339391593014],
-        [0.10117257181649406, 0.08608351243763912, 0.24075089154601287],
[…]
