### 2020-10-19 #2629 @amanchokshi
@mbobra I've fixed the bug. If you update embers to the latest version 0.7.3, the examples should now work as shown in the docs
