### 2020-10-19 #2629 @mbobra
I'm sorry -- I'm probably missing something simple. I'm looking at the [Embers by Example](https://embers.readthedocs.io/en/latest/embersbyexample.html) guide, which says:

```
$ mkdir ~/EMBERS
$ cd ~/EMBERS
```

So I did that. Then I skipped the stuff about the test tool. Then the next code block says:

```python
waterfall_single
```

But I'm still confused 😟 Where are the sample data? How do I access these data? What import statements do I need to make before I can run `waterfall_single`? Are there any example notebooks I can look at?

I was looking at a recently published JOSS paper called [PyCS3](https://joss.theoj.org/papers/10.21105/joss.02654) (which I picked at random). They provide a [tutorial](https://cosmograil.gitlab.io/PyCS3/tutorial/generalities.html) and [example notebooks](https://gitlab.com/cosmograil/PyCS3/-/tree/master/notebook), which users can download and run. I don't want to be overly prescriptive -- so PYCS3's approach is by no means the only way to provide examples -- but I personally find it easy to understand. Perhaps this can be fixed by a few explanatory sentences if I am missing something simple.
