### 2020-10-20 48a27bb
Added link to api docs in embersbyexample
