### 2020-10-20 #2629 @amanchokshi
@mbobra

`waterfall_single` is a cli tool which makes it easy to interact with the underlying [single_waterfall](https://embers.readthedocs.io/en/latest/api-rf_tools.html#embers.rf_tools.rf_data.single_waterfall) function.

Internally, what `waterfall_single` does:
```
import pkg_resources
from embers.rf_tools.rf_data import single_waterfall

# If no data provided, use package sample data
rf_file = pkg_resources.resource_filename("embers.kindle", "data/rf_data/S06XX/2019-10-10/S06XX_2019-10-10-02:30.txt")
out_dir  = "embers_out/rf_tools"

single_waterfall(rf_file, out_dir)
```

In [Embers by Example](https://embers.readthedocs.io/en/latest/embersbyexample.html#rf-tools) I have added a link which takes you to the api docs of the relevant function. If you think this is sufficient, I can all similar links to all the examples.

Alternately, I could create a notebook, with accompanying example data.
