### 2020-10-24 bf688e5
Moved waterfall_batch function to rf_tools.rf_data. Fixed docs to match move
