### 2020-10-24 8cfaf86
Moved align_batch to embers.rf_tools.align_data. Modified docs to reflect change
