+++ docs/api.rst
@@ -27,6 +27,7 @@ Contains :mod:`~embers.rf_tools.rf_data`, :mod:`~embers.rf_tools.align_data`, :m
     align_data.savgol_interp
     align_data.plot_savgol_interp
     align_data.save_aligned
+    align_data.align_batch
     colormaps.spectral
     colormaps.jade
     colormaps.waves_2d
