+++ docs/embersbyexample.rst
@@ -68,8 +68,8 @@ RF Tools
 
 Waterfall Plots
 ^^^^^^^^^^^^^^^
-To get a quick preview of the raw RF data, we create waterfall plots. Creates a waterfall plot of sample data provided with *EMBERS* using 
-the :func:`~embers.rf_tools.rf_data.single_waterfall` function with the :samp:`waterfall_single` cli tool: 
+To get a quick preview of the raw RF data, we create waterfall plots. Creates a waterfall plot of sample data provided with *EMBERS* using
+the :func:`~embers.rf_tools.rf_data.single_waterfall` function with the :samp:`waterfall_single` cli tool:
 
 .. code-block:: console
 
@@ -138,13 +138,14 @@ To get a preview of how amazing they are
 Align Data
 ^^^^^^^^^^
 
-The RF Explorers used to record satellite data did not record data at exactly the same frequency and do not start recording at exactly the same time.
+The RF Explorers used to record satellite data may not record data at exactly the same frequency and may not start recording at exactly the same time.
 In fact, the older models record at approximately 6 Hz, while the newer ones are capable of a sampling rate of nearly 9 Hz. This discrepency in sampling
 rates makes it difficult to compare any two data samples. This issue is overcome by smoothing the data, along the time axis, with a Savitzky-Golay filter.
-Interpolating the smoothed data and resampling it at a constant frequency [ 0.5 Hz ] gives us a easier data set to work with.
+Interpolating the smoothed data and resampling it at a constant frequency [ 1 Hz ] gives us a easier data set to work with.
 
-Two level of savgol filters are applied, first to capture deep nulls + small structure, and second level to smooth over noise. :samp:`align_single` can be
-used to play with the various parameters available. Sensible defaults are provided as a starting point. The following code plots one frequency channel of
+Two level of savgol filters are applied, first to capture deep nulls + small structure, and second level to smooth over noise. A cli tool :samp:`align_single`,
+based on the :func:`~embers.rf_tools.align_data.plot_savgol_interp` function,
+can be used to play with the various parameters available. Sensible defaults are provided as a starting point. The following code plots one frequency channel of
 RF data and shows the efficacy of the selected smoothing filter.
 
 .. code-block:: console
@@ -160,18 +161,90 @@ RF data and shows the efficacy of the selected smoothing filter.
     Saving sample savgol_interp plot to: ./embers_out/rf_tools
 
 
+Alternately, the following sample code may be used to achieve identical results:
+
+.. code-block:: python
+
+    import pkg_resources
+    from embers.rf_tools.align_data import plot_savgol_interp
+
+    savgol_window_1=11
+    savgol_window_2=15
+    polyorder=2
+    interp_type="cubic"
+    interp_freq=1
+    channel=59
+    out_dir="./embers_out/rf_tools"
+
+
+    ref_file = pkg_resources.resource_filename(
+        "embers.kindle", "data/rf_data/rf0XX/2019-10-10/rf0XX_2019-10-10-02:30.txt"
+    )
+
+    tile_file = pkg_resources.resource_filename(
+        "embers.kindle", "data/rf_data/S06XX/2019-10-10/S06XX_2019-10-10-02:30.txt"
+    )
+
+
+    plot_savgol_interp(
+        ref=ref_file,
+        tile=tile_file,
+        savgol_window_1=savgol_window_1,
+        savgol_window_2=savgol_window_2,
+        polyorder=polyorder,
+        interp_type=interp_type,
+        interp_freq=interp_freq,
+        channel=channel,
+        out_dir=out_dir,
+    )
+
+
 .. image:: _static/imgs/align_data.png
     :width: 100%
     :alt: EMBERS custom colormaps
 
[…]
