### 2020-10-24 1f0b8b3
Sample colormaps code-block
