### 2020-10-25 aa6ebcb
Made axis label font size larger for null test
