### 2020-10-25 3f97296
Moved ephem_batch function to embers.sat_utils.sat_ephemeris
