+++ docs/api-sat_utils.rst
@@ -17,6 +17,7 @@
 .. autofunction:: embers.sat_utils.sat_ephemeris.ephem_data
 .. autofunction:: embers.sat_utils.sat_ephemeris.sat_plot
 .. autofunction:: embers.sat_utils.sat_ephemeris.save_ephem
+.. autofunction:: embers.sat_utils.sat_ephemeris.ephem_batch
 
 .. automodule:: embers.sat_utils.chrono_ephem
 .. autofunction:: embers.sat_utils.chrono_ephem.obs_times
