### 2020-10-25 368c57c
Example batch_ephem code
