+++ src/embers/kindle/ephem_chrono.py
@@ -3,7 +3,7 @@ Chronological Ephemeris
 -----------------------
 
 Collate ephemeris data generated by :mod:`~embers.sat_utils.sat_ephemeris` for multiple satellites
-and determine all ephemeris present in 30 minute observation windows. Save chronological ephemeris 
+and determine all ephemeris present in 30 minute observation windows. Save chronological ephemeris
 data to json files in :samp:`./embers_out/sat_utils/ephem_chrono`
 
 """
@@ -78,7 +78,7 @@ def main():
     """Execute chrono_ephem from terminal."""
 
     print(f"Saving chronological Ephem files to: {_out_dir}")
-    print(f"Grab a coffee, this may take more than a couple of minutes!")
+    print("Grab a coffee, this may take more than a couple of minutes!")
 
     # save log file
     Path(_out_dir).mkdir(parents=True, exist_ok=True)
