+++ docs/embersbyexample.rst
@@ -374,12 +374,39 @@ or with the equivalent example script
 
 Chronological ephemeris
 ^^^^^^^^^^^^^^^^^^^^^^^
-Collate ephemeris data generated by :samp:`ephem_batch` for multiple satellites and determine all satellites present in each 30 minute observation and what their
-trajectories at the geographic location. Save chronological ephemeris data to json files in :samp:`./embers_out/sat_utils/ephem_chrono`.
+Collate ephemeris data generated above by :samp:`ephem_single` or :samp:`ephem_batch` for multiple satellites and determine all satellites present in each
+30 minute observation and what their trajectories at the geographic location. The :func:`~embers.sat_utils.chrono_ephem.save_chrono_ephem` function saves
+chronological ephemeris data to json files in :samp:`./embers_out/sat_utils/ephem_chrono`.
+
+Use the following cli tool to collate satellite data
 
 .. code-block:: console
 
     $ ephem_chrono --start_date=YYYY-MM-DD --stop_date=YYYY-MM-DD
+
+or the equivalent sample script
+
+.. code-block:: python
+
+    from embers.sat_utils.chrono_ephem import save_chrono_ephem
+
+    time_zone = "Australia/Perth"
+    start_date = "2019-10-01"
+    stop_date = "2019-10-10"
+    interp_type = "cubic"
+    interp_freq = 1
+    ephem_dir = "./embers_out/sat_utils/ephem_data"
+    out_dir = "./embers_out/sat_utils/ephem_chrono"
+
+    save_chrono_ephem(
+        time_zone,
+        start_date,
+        stop_date,
+        interp_type,
+        interp_freq,
+        ephem_dir,
+        out_dir,
+    )
 
 Satellite Channels
 ^^^^^^^^^^^^^^^^^^
