+++ src/embers/kindle/data/TLE/25417.txt
@@ -1,0 +1,28 @@
+1 25417U 98046E   19273.52295833  .00000044  00000-0  66616-4 0  9992
+2 25417  44.9959 293.3827 0005543 289.3929  70.6318 14.33127639104823
+1 25417U 98046E   19273.87150887 +.00000044 +00000-0 +66573-4 0  9991
+2 25417 044.9959 291.7429 0005543 290.9384 069.0869 14.33127761104702
+1 25417U 98046E   19274.56860990  .00000043  00000-0  66461-4 0  9996
+2 25417  44.9957 288.4635 0005524 293.9855  66.0413 14.33128027104974
+1 25417U 98046E   19274.84745028 +.00000043 +00000-0 +66349-4 0  9996
+2 25417 044.9957 287.1517 0005523 295.2048 064.8226 14.33128129104844
+1 25417U 98046E   19275.61426139  .00000024  00000-0  59543-4 0  9998
+2 25417  44.9956 283.5443 0005505 298.6648  61.3644 14.33128330105126
+1 25417U 98046E   19276.79933299 -.00000009  00000-0  47491-4 0  9991
+2 25417  44.9955 277.9693 0005479 304.0545  55.9781 14.33128596105290
+1 25417U 98046E   19276.86904308 -.00000007 +00000-0 +48117-4 0  9992
+2 25417 044.9955 277.6413 0005477 304.3486 055.6841 14.33128619105135
+1 25417U 98046E   19277.84498429 -.00000025  00000-0  41603-4 0  9996
+2 25417  44.9954 273.0500 0005444 308.6901  51.3458 14.33128803105440
+1 25417U 98046E   19277.91469437 -.00000023 +00000-0 +42366-4 0  9996
+2 25417 044.9954 272.7220 0005441 308.9888 051.0473 14.33128823105287
+1 25417U 98046E   19278.89063548 -.00000026 +00000-0 +41343-4 0  9994
+2 25417 044.9954 268.1307 0005399 313.3071 046.7325 14.33129023105424
+1 25417U 98046E   19279.16947575 -.00000028  00000-0  40810-4 0  9996
+2 25417  44.9954 266.8188 0005377 314.4853  45.5553 14.33129085105633
+1 25417U 98046E   19279.93628655 -.00000034 +00000-0 +38400-4 0  9995
+2 25417 044.9954 263.2113 0005340 317.8335 042.2100 14.33129169105572
+1 25417U 98046E   19280.49396717 -.00000048  00000-0  33632-4 0  9996
+2 25417  44.9954 260.5878 0005311 320.4513  39.5945 14.33129196105831
+1 25417U 98046E   19280.91222761 -.00000048 +00000-0 +33380-4 0  9990
+2 25417 044.9954 258.6201 0005293 322.3872 037.6603 14.33129214105715
