+++ src/embers/kindle/data/TLE/44387.txt
@@ -1,104 +1,0 @@
-1 44387U 19038A   19273.18553806 -.00000035  00000-0  33891-5 0  9993
-2 44387  98.5797 234.0152 0001713 151.4465 208.6808 14.23334367 12368
-1 44387U 19038A   19273.46672789 -.00000037  00000-0  26646-5 0  9998
-2 44387  98.5797 234.2894 0001728 150.8774 209.2502 14.23334372 12404
-1 44387U 19038A   19273.81821518 -.00000037  00000-0  25491-5 0  9990
-2 44387  98.5798 234.6321 0001748 150.2434 209.8844 14.23334374 12455
-1 44387U 19038A   19273.88851264 -.00000041 +00000-0 +85552-6 0  9998
-2 44387 098.5798 234.7007 0001751 150.0797 210.0482 14.23334369012468
-1 44387U 19038A   19274.09940502 -.00000040  00000-0  10857-5 0  9990
-2 44387  98.5799 234.9063 0001767 149.6404 210.4877 14.23334364 12499
-1 44387U 19038A   19274.52118974 -.00000037  00000-0  26217-5 0  9991
-2 44387  98.5799 235.3176 0001787 148.7845 211.3440 14.23334387 12558
-1 44387U 19038A   19274.80237954 -.00000037  00000-0  25141-5 0  9997
-2 44387  98.5800 235.5918 0001794 148.3918 211.7368 14.23334397 12590
-1 44387U 19038A   19274.87267700 -.00000036 +00000-0 +28702-5 0  9991
-2 44387 098.5800 235.6604 0001799 148.2742 211.8545 14.23334405012602
-1 44387U 19038A   19275.22416426 -.00000028  00000-0  65116-5 0  9993
-2 44387  98.5801 236.0031 0001823 147.4072 212.7220 14.23334438 12658
-1 44387U 19038A   19275.43505663 -.00000028  00000-0  65018-5 0  9999
-2 44387  98.5801 236.2088 0001830 147.0311 213.0982 14.23334441 12688
-1 44387U 19038A   19275.50535408 -.00000028  00000-0  65286-5 0  9994
-2 44387  98.5802 236.2773 0001834 146.8708 213.2586 14.23334435 12697
-1 44387U 19038A   19275.71624645 -.00000028  00000-0  66286-5 0  9990
-2 44387  98.5802 236.4830 0001845 146.4983 213.6313 14.23334428 12724
-1 44387U 19038A   19276.34892355 -.00000024  00000-0  81934-5 0  9998
-2 44387  98.5804 237.0999 0001879 145.2574 214.8728 14.23334457 12814
-1 44387U 19038A   19276.77070831 -.00000024  00000-0  83393-5 0  9993
-2 44387  98.5804 237.5112 0001896 144.2970 215.8335 14.23334445 12872
-1 44387U 19038A   19276.84100577 -.00000024  00000-0  82048-5 0  9998
-2 44387  98.5804 237.5798 0001899 144.1738 215.9568 14.23334449 12880
-1 44387U 19038A   19276.91130322 -.00000023 +00000-0 +89433-5 0  9991
-2 44387 098.5805 237.6483 0001903 144.0629 216.0678 14.23334459012893
-1 44387U 19038A   19277.19249306 -.00000022  00000-0  90414-5 0  9995
-2 44387  98.5805 237.9225 0001919 143.3590 216.7720 14.23334453 12930
-1 44387U 19038A   19277.47368289 -.00000022  00000-0  91912-5 0  9992
-2 44387  98.5806 238.1967 0001935 142.8081 217.3232 14.23334463 12974
[…]
