### 2020-10-25 3693ac3
Trying to catch a weird excwption
