### 2020-10-25 b6dfc43
Trying to catch a weird excwption
