### 2020-10-25 d3d42e2
Trying to catch a weird excwption
