### 2020-10-25 f4aaf23
Test_window_chan_map_err is obsolete. Changed code to deal with that err better
