### 2020-10-25 99d0f08
Test_window_chan_map_err is obsolete + nuked
