### 2020-10-26 23e345f
Used sensible defaluts if no arguments are provided
