### 2020-10-26 e468fa8
Used sensible defaluts if no arguments are provided
