### 2020-10-26 01ac4f3
cleaned up sample code for mwa_pointings
