### 2020-10-26 29b032b
Cleaned up intro. Added sample data download instructions
