### 2020-10-26 e9b6e4d
Waterfall batch with external example data
