+++ src/embers/kindle/waterfall_batch.py
@@ -11,7 +11,6 @@ import argparse
 import logging
 from pathlib import Path
 
-import pkg_resources
 from embers.rf_tools.rf_data import waterfall_batch
 
 _parser = argparse.ArgumentParser(
@@ -21,15 +20,24 @@ _parser = argparse.ArgumentParser(
 )
 
 _parser.add_argument(
-    "--start_date", metavar="\b", default="", help="start date in YYYY-MM-DD format"
+    "--start_date",
+    metavar="\b",
+    default="2019-10-10",
+    help="start date in YYYY-MM-DD format, default=2019-10-10",
 )
 
 _parser.add_argument(
-    "--stop_date", metavar="\b", default="", help="stop date in YYYY-MM-DD format"
+    "--stop_date",
+    metavar="\b",
+    default="2019-10-10",
+    help="stop date in YYYY-MM-DD format, default=2019-10-10",
 )
 
 _parser.add_argument(
-    "--data_dir", metavar="\b", default="", help="root of dir where rf data is saved"
+    "--data_dir",
+    metavar="\b",
+    default="./tiles_data",
+    help="root of dir where rf data is saved, default=tiles_data",
 )
 
 _parser.add_argument(
@@ -44,24 +52,6 @@ _start_date = _args.start_date
 _stop_date = _args.stop_date
 _data_dir = _args.data_dir
 _out_dir = _args.out_dir
-
-# if no input file provided, use sample package data
-if _data_dir == "":
-    print("----------------------------------------------------------")
-    print("No data dir provided, using packaged sample data")
-    _data_dir = pkg_resources.resource_filename("embers.kindle", "data/rf_data/")
-
-if _start_date == "":
-    print("No start_date provided, using 2019-10-01 for sample data")
-    _start_date = "2019-10-01"
-
-if _stop_date == "":
-    print("No stop_date provided, using 2019-10-10 for sample data")
-    print("")
-    print(">>> waterfall_batch --help, for more options")
-    print("----------------------------------------------------------")
-    _stop_date = "2019-10-10"
-
 
 # Logging config
 _log_dir = Path(f"{_out_dir}/waterfalls")
