### 2020-10-26 3714416
Clean terminal code-block
