### 2020-10-26 97141d0
Align batch with external example data
