### 2020-10-26 e065bda
ephem_batch with external example data
