+++ docs/embersbyexample.rst
@@ -10,7 +10,7 @@ Setup
 -----
 *EMBERS* works best in a virtual environment and requires Python > 3.6. Follow the :doc:`Installation Instructions <installation>` to setup a suitable enviroment.
 
-.. code-block:: console
+.. code-block::
 
     $ mkdir ~/EMBERS
     $ cd ~/EMBERS
@@ -19,7 +19,7 @@ The above creates an :samp:`EMBERS` directory, within which we can run our examp
 
 We now download sample data required to run the following examples
 
-.. code-block:: bash
+.. code-block::
 
     git clone https://github.com/amanchokshi/EMBERS-Sample-Data.git
     mv EMBERS-Sample-Data/tiles_data .
@@ -40,7 +40,7 @@ sub directories for each MWA and reference tile exist (:samp:`S06XX`, :samp:`S06
 there exists a date directory for every day of observations in the :samp:`YYYY-MM-DD` format, within which live the raw RF data, in binary :samp:`.txt` files
 saved every 30 minutes, with the naming convention as follows :samp:`S06XX_YYYY-DD-MM-hh:mm.txt`
 
-.. code-block:: console
+.. code-block::
 
     tiles_data
     ├── S06XX
@@ -72,7 +72,7 @@ Waterfall Plots
 To get a quick preview of the raw RF data, we create waterfall plots. Creates a waterfall plot of sample data provided with *EMBERS* using
 the :func:`~embers.rf_tools.rf_data.single_waterfall` function with the :samp:`waterfall_single` cli tool:
 
-.. code-block:: console
+.. code-block::
 
     $ waterfall_single
     >>> Waterfall plot saved to ./embers_out/rf_tools/S06XX_2019-10-10-02:30.png
@@ -121,7 +121,7 @@ Colormaps
 visualise raw data and maximize dynamic range, while :samp:`jade` is perceptually uniform and sequential and is suitable for science.
 To get a preview of how amazing they are
 
-.. code-block:: console
+.. code-block::
 
     $ colormaps
 
@@ -153,7 +153,7 @@ based on the :func:`~embers.rf_tools.align_data.plot_savgol_interp` function,
 can be used to play with the various parameters available. Sensible defaults are provided as a starting point. The following code plots one frequency channel of
 RF data and shows the efficacy of the selected smoothing filter.
 
-.. code-block:: console
+.. code-block::
 
     $ align_single
     >>> Saving sample savgol_interp plot to: ./embers_out/rf_tools
@@ -200,7 +200,7 @@ MWA tile are smoothed and aligned and saved to compressed :samp:`npz` file by :f
 
 The :samp:`align_batch` cli tool is a convenient way to align large volumes of data
 
-.. code-block:: console
+.. code-block::
 
     $ align_batch
     >>> Aligned files saved to: ./embers_out/rf_tools/align_data
@@ -251,7 +251,7 @@ Download Ephemeris
 
 Once valid login credentials have been obtained, download tle files with the :func:`~embers.sat_utils.sat_list.download_tle` using the following cli tool
 
-.. code-block:: console
+.. code-block::
 
     $ download_tle --start_date=YYYY-MM-DD --stop_date=YYYY-MM-DD --st_ident=** --st_pass=**
 
@@ -320,7 +320,7 @@ Analysing a batch of TLE files is achieved with the :func:`embers.sat_utils.sat_
 :samp:`download_tle` into satellite ephemeris data: rise time, set time, alt/az arrays at a given time cadence. This is saved to a npz file which will be used
 to plot the satellite sky coverage over the geographic location supplied. It can be used with the following cli tool
 
-.. code-block:: console
+.. code-block::
 
     $ ephem_batch
[…]
