### 2020-10-26 66eb273
sat_channels with external example data
