### 2020-10-26 afd017b
sat_channels with external example data
