### 2020-10-26 ba1eca0
sat_channels with external example data
