### 2020-10-26 aa47f57
mwa_pointings with external example data
