### 2020-10-26 ebc44ec
mwa_pointings with external example data
