### 2020-10-26 c1146a5
mwa_pointings with external example data
