### 2020-10-26 2eb586b
mwa_pointings with external example data
