+++ docs/embersbyexample.rst
@@ -510,14 +510,12 @@ either the following cli tool or the sample script
 
 .. code-block:: python
 
-    import pkg_resources
-    from pathlib import Path
     from embers.mwa_utils.mwa_pointings import mwa_point_meta
 
-    start_date = "2019-10-10"
-    stop_date = "2019-10-10"
-    num_pages = 15
-    time_thresh = 5
+    start_date = "2019-10-09"
+    stop_date = "2019-10-11"
+    num_pages = 4
+    time_thresh = 1
     time_zone = "Australia/Perth"
     rf_dir = "./tiles_data"
     out_dir = "./embers_out/mwa_utils"
