### 2020-10-26 2d74366
mwa_dipoles with external example data
