### 2020-10-26 16a8e61
mwa_dipoles with external example data
