### 2020-10-26 aad2ab8
mwa_fee with external example data
