### 2020-10-26 6b40ac4
mwa_fee with external example data
