### 2020-10-26 bb91266
ref_models with external example data
