### 2020-10-26 9bc7946
Trying to fix binned_stats nan problem, maybe?
