### 2020-10-26 bd6d8c1
Trying to fix binned_stats nan problem, maybe?
