+++ src/embers/tile_maps/tile_maps.py
@@ -638,7 +638,8 @@ def rfe_calibration(
                                                         # consider residuals of sats which pass within 10 deg of zenith
                                                         # an hp index of 111 approx corresponds to a zenith angle of 10 degrees
                                                         hp_10_deg = 111
-
+                                                        
+                                                        print(np.amin(u))
                                                         if np.amin(u) <= hp_10_deg:
 
                                                             # only passes longer than 10 minutes
