+++ src/embers/kindle/compare_beams.py
@@ -53,7 +53,4 @@ def main():
     """Execute null test from terminal."""
     batch_compare_beam(_nside, _fee_map, _map_dir, _out_dir)
 
-    print(
-        "-----------------------------------------------------------------"
-    )
-    print(f"Beam comparison plots saved to {_out_dir}")
+    print(f"Beam comparison plots saved to: {_out_dir}")
