### 2020-10-26 #2629 @amanchokshi
Hi @mbobra, there have been quite significant updates to the [Embers by Example](https://embers.readthedocs.io/en/latest/embersbyexample.html) section of the documentation.

Can you begin by updating to version 0.7.4 and deleting all the old files you may have created?

The instructions begin with downloading one day of sample data from a new data repository, which all the subsequent examples use. Each section now has instruction on how to use the cli-tool, with an equivalent example code-block below. I've also linked the functions use which will take you to the relevant section of the [EMBERS API](https://embers.readthedocs.io/en/latest/api.html).

Certain examples will take a little while to run and I expect that you can run all the examples in less than 3 hours. It's important to run them in the sequence presented, because data products created early on are used at later stages.

Hope things work this time and are much more clear 😁
