### 2020-10-26 #2629 @teuben
Just a suggestion on this test though:   is there no way you could make
this test run in a few minutes?  If a cosmological simulation takes 3
hours, you can either take fewer steps or a smaller grid, but it should
functionally give some answer (that can be compared to something given
as baseline result).  I'm not too happy that I would need to wait 3
hours on a test.
