### 2020-10-26 36f842a
Updated JOSS paper to remove Chokshi et al in prep
