### 2020-10-27 #2629 @amanchokshi
Hi @teuben, the tests for embers are automated, and take ~20 minutes on Travis.

The examples shown in [Embers by Example](https://embers.readthedocs.io/en/latest/embersbyexample.html) do take longer, but are meant to be a guide to show new users how to begin interacting with Embers. I've reduces the amount of sample data as much as possible (only 1 day of data, instead of 6 months), but some of the later functions are statistical in nature and require multiple satellite passes along with their corresponding data files to accurately calibrate the instrument.
