+++ docs/embersbyexample.rst
@@ -72,7 +72,7 @@ Waterfall Plots
 To get a quick preview of the raw RF data, we create waterfall plots. Creates a waterfall plot of sample data provided with *EMBERS* using
 the :func:`~embers.rf_tools.rf_data.single_waterfall` function with the :samp:`waterfall_single` cli tool:
 
-.. code-block::
+.. code-block:: console
 
     $ waterfall_single
     >>> Waterfall plot saved to ./embers_out/rf_tools/S06XX_2019-10-10-02:30.png
@@ -96,7 +96,7 @@ Or using :func:`~embers.rf_tools.rf_data.single_waterfall` as shown in the examp
 We can also create a set of waterfall plots for all rf_files within a date interval using the :func:`~embers.rf_tools.rf_data.waterfall_batch` function, with
 either the provided cli tool or with the following example code
 
-.. code-block::
+.. code-block:: console
 
     $ waterfall_batch
     >>> Processing rf data files between 2019-10-10 and 2019-10-10
@@ -121,7 +121,7 @@ Colormaps
 visualise raw data and maximize dynamic range, while :samp:`jade` is perceptually uniform and sequential and is suitable for science.
 To get a preview of how amazing they are
 
-.. code-block::
+.. code-block:: console
 
     $ colormaps
 
@@ -153,7 +153,7 @@ based on the :func:`~embers.rf_tools.align_data.plot_savgol_interp` function,
 can be used to play with the various parameters available. Sensible defaults are provided as a starting point. The following code plots one frequency channel of
 RF data and shows the efficacy of the selected smoothing filter.
 
-.. code-block::
+.. code-block:: console
 
     $ align_single
     >>> Saving sample savgol_interp plot to: ./embers_out/rf_tools
@@ -200,7 +200,7 @@ MWA tile are smoothed and aligned and saved to compressed :samp:`npz` file by :f
 
 The :samp:`align_batch` cli tool is a convenient way to align large volumes of data
 
-.. code-block::
+.. code-block:: console
 
     $ align_batch
     >>> Aligned files saved to: ./embers_out/rf_tools/align_data
@@ -251,7 +251,7 @@ Download Ephemeris
 
 Once valid login credentials have been obtained, download tle files with the :func:`~embers.sat_utils.sat_list.download_tle` using the following cli tool
 
-.. code-block::
+.. code-block:: console
 
     $ download_tle --start_date=YYYY-MM-DD --stop_date=YYYY-MM-DD --st_ident=** --st_pass=**
 
@@ -290,7 +290,7 @@ the computation of satellites trajectories over a geographical location (MWA tel
 the :func:`~embers.sat_utils.sat_ephemeris.save_ephem` with either the following cli tool or the equivalent sample code.
 
 
-.. code-block::
+.. code-block:: console
 
     $ ephem_single
     >>> Saved sky coverage plot of satellite [25417] to ./embers_out/sat_utils/ephem_plots
@@ -320,7 +320,7 @@ Analysing a batch of TLE files is achieved with the :func:`embers.sat_utils.sat_
 :samp:`download_tle` into satellite ephemeris data: rise time, set time, alt/az arrays at a given time cadence. This is saved to a npz file which will be used
 to plot the satellite sky coverage over the geographic location supplied. It can be used with the following cli tool
 
-.. code-block::
+.. code-block:: console
 
     $ ephem_batch
     >>> Saving logs to ./embers_out/sat_utils/ephem_data
@@ -351,7 +351,7 @@ chronological ephemeris data to json files in :samp:`./embers_out/sat_utils/ephe
 
 Use the following cli tool to collate satellite data
 
-.. code-block::
+.. code-block:: console
 
     $ ephem_chrono
[…]
