### 2020-10-29 4d0c145
Moving to v0.7.7 and trying to add option to turn off rfe_cali
