### 2020-10-30 7f1829d
Fixed MWA fee dipole flagging and moving to v0.7.8
