+++ src/embers/kindle/mwa_fee.py
@@ -34,7 +34,7 @@ _parser.add_argument(
 _parser.add_argument(
     "--flags",
     metavar="\b",
-    default="0",
+    default='0',
     type=str,
     help="List of flagged dipoles with indices from 1 to 32. 1-16 are dipoles of XX pol while 17-32 are for YY. Ex: flags='1,17' represents the first dipole of the XX & YY tiles as being flagged and having a amplitude of 0. Default=[]"
 )
@@ -49,7 +49,13 @@ _parser.add_argument(
 _args = _parser.parse_args()
 _nside = _args.nside
 _pointings = [int(item) for item in _args.pointings.split(',')]
-_flags = [int(item) for item in _args.flags.split(',') if not "0"]
+
+_flags = [int(item) for item in _args.flags.split(',')]
+
+if _flags == [0]:
+    _flags = []
+
+print(_flags)
 _out_dir = _args.out_dir
 
 
