### 2020-11-15 adc6973
Fixed legend on rfe_gain plot
