### 2020-11-15 c4bb583
Fixed legend on rfe_gain plot
