### 2020-11-15 770f072
Moving to V0.7.9 and fixed rfe_gain legend
