### 2020-11-15 d887171
Removing unnecessary python versions from travis
