+++ .travis.yml
@@ -5,11 +5,7 @@ jobs:
   include:
     - python: 3.6
     - python: 3.7
-    - python: 3.7-dev
     - python: 3.8
-    - python: 3.8-dev
-    - python: 3.9-dev
-  allow_failures:
     - python: 3.9-dev
 
 before_install:
