### 2020-11-15 35cdcbc
Added h5py to reqs. Hopefully fixes travis break
