+++ requirements.txt
@@ -3,6 +3,7 @@ certifi==2020.4.5.2
 chardet==3.0.4
 cycler==0.10.0
 healpy==1.13.0
+h5py==2.10.0
 idna==2.9
 jplephem==2.14
 kiwisolver==1.2.0
