### 2020-11-16 357c276
Issue 6. Added working about example data size
