+++ docs/embersbyexample.rst
@@ -17,7 +17,8 @@ Setup
 
 The above creates an :samp:`EMBERS` directory, within which we can run our example code. Outputs will by default, be saved to :samp:`~/EMBERS/embers_out`.
 
-We now download sample data required to run the following examples
+We now download sample data required to run the subsequent examples from the `EMBERS Sample Data github repository <https://github.com/amanchokshi/EMBERS-Sample-Data>`_.
+This repository contains one day of satellite radio data and will require 4GB of disk space.
 
 .. code-block:: console
 
