### 2020-11-16 ca79ec9
Issue 6. Added WARNING about example data size
