### 2020-11-16 d67fe6c
Issue 6. Clarify output directory relative location
