### 2020-11-16 0f16610
Added --depth option to cloning sample data
