### 2020-11-16 #2629 @teuben
I've been adding some issues, author is working on those. I'm about halfway through the examples, a bit more work than I thought, will take a night sleep on it. more tomorrow.
