### 2020-11-16 609c415
Restructured so that sphinx docs can be made
