### 2020-11-16 8272d67
Restructured so that sphinx docs can be made
