+++ src/embers/kindle/__init__.py
@@ -1,4 +1,3 @@
 """
-:mod:`embers.kindle` contains a set of executable script to process data in various ways
-
+:mod:`embers.kindle` contains a set of command-line (cli) tools or executable script to process data in various ways
 """
