### 2020-11-16 d296fd7
Removed unnecessary heading in doc-string
