+++ docs/api-kindle.rst
@@ -5,4 +5,4 @@
 
 .. automodule:: embers.kindle
 
-
+.. autofunction:: embers.kindle.waterfall_single.main
