### 2020-11-16 c14d224
Documented waterfall_batch
