+++ src/embers/kindle/waterfall_single.py
@@ -42,7 +42,5 @@ def main():
     _rf_file = _args.rf_file
     _out_dir = _args.out_dir
 
-    """Executable: save single waterfall plot"""
-
     print(f"Waterfall plot saved to ./{_out_dir}/{Path(_rf_file).stem}.png")
     single_waterfall(_rf_file, _out_dir)
