+++ docs/api-kindle.rst
@@ -6,3 +6,4 @@
 .. automodule:: embers.kindle
 
 .. autofunction:: embers.kindle.waterfall_single.main
+.. autofunction:: embers.kindle.waterfall_batch.main
