### 2020-11-16 1ec8e30
Added documentation for colormaps
