+++ docs/api-kindle.rst
@@ -7,3 +7,4 @@
 
 .. autofunction:: embers.kindle.waterfall_single.main
 .. autofunction:: embers.kindle.waterfall_batch.main
+.. autofunction:: embers.kindle.colormaps.main
