### 2020-11-16 c530df0
API documentation for embers.kindle is complete
