### 2020-11-17 2d8a422
Added+fixed print statements pointing to output files
