### 2020-11-17 f7895fc
Added+fixed print statements pointing to output files
