### 2020-11-17 40cf300
Added+fixed print statements pointing to output files
