### 2020-11-17 713f3b4
Moving to v0.8.0 to fix cli-tools printing path to output file
