### 2020-11-17 #2629 @teuben
@amanchokshi  apart from some minor things I know you are looking into, my last hurdle on the checkmark list is performance.

There are no good comments on performance that I could find (in fact, in the Example there should be some warnings in more places, but you are working on the aggressive CPU usage already). I have not looked at the code, but let me try to explain how I would lke to try and work towards an answer why this is so slow.   We have N files, with each file M lines. Yet, the example you mentioned took 40 core hours.  Can this be understood in perhaps an O(N) analysis?  Are there no loops  in python?  Is everything done in good numpy notation,without looping?
There is of course the potential speedup using XLA/numba/dask, which is mentioned in one of the issues, but before resorting to this, would be nice if we could understand in simple terms why it takes as long as it does.
