### 2020-11-17 83c108b
Added --max_cores option to limit number of cores availible to parallel processes
