### 2020-11-17 9c6727d
Added --max_cores option to limit number of cores availible to parallel processes
