### 2020-11-17 6b96af2
Moving to V0.8.1 with option to limit number of cores for paralelized cli-tools
