### 2020-11-17 51e31b0
Added instructions to run automated tests
