### 2020-11-17 #2629 @teuben
> I've added instructions to run the automated tests at the bottom of the examples page - [Testing EMBERS](https://embers.readthedocs.io/en/latest/embersbyexample.html#testing-embers). It should take no more than 2 minutes to run the tests. To run the test you need to clone the directory and install a few additional dependancies:
>
> ```
> git clone https://github.com/amanchokshi/EMBERS.git
> cd EMBERS
>
> # Setup a virtual enviroment
> python -m venv embers-env
> source embers-env/bin/activate
>
> pip install .
> pip install -r requirements_dev.txt
>
> # Run automated tests
> pytest
> ```

this ran fine for me.   I had two failures (and a few deprecation warnings):

      tests/test_mwa_utils/test_mwa_pointings.py::test_clean_meta_json_start FAILED
      tests/test_tile_maps/test_ref_fee_healpix.py::test_create_model_last FAILED
