### 2020-11-18 #2629 @amanchokshi
Regarding the performance, I've re-run `align_batch` using only a single core, with a varying number of files. The probes your question of how the problem scales with N. The results seem to indicate a linear relationship between number of files and processing time.

![files_time](https://user-images.githubusercontent.com/40157844/99484385-65944500-29b4-11eb-8c82-0fb29f18ccb9.png)

I'm quite out of my depth with this kind of code testing, but let me try and clarify exactly what the code does. Each data file has ~17k lines, which each contain 112 floating point numbers corresponding to the 112 frequency channels available. The 17k lines correspond to 17k timesteps at which power was samples.  A pair of two such files are selected.

- The two arrays are interpolated down to a lower frequency (1Hz default). This reduces the array to (1800, 112)
- A two stage [savgol](https://docs.scipy.org/doc/scipy/reference/generated/scipy.signal.savgol_filter.html) filter is applied along the time axis (1800 points).  This is a 1D filter along the longer axis of the array.
- The two interpolated and smoothed arrays are saved to disk
