### 2020-11-19 #2629 @teuben
> The depreciation warnings come from a package which I use - mwa-pb, and has required me to freeze some of the dependancies in the requirement.txt. I'm not sure why those two tests are failing. I cloned a fresh copy of EMBERS and tried the tests again and all of them seem to pass. Could it be a python version thing? I'm stumped
>
> ![Screen Shot 2020-11-18 at 1 13 39 pm](https://user-images.githubusercontent.com/40157844/99474431-ad10d600-29a0-11eb-8b21-f162a801658c.png)
> ![Screen Shot 2020-11-18 at 1 13 06 pm](https://user-images.githubusercontent.com/40157844/99474440-b0a45d00-29a0-11eb-98bd-bb4f36a82a9f.png)
>
> The automated tests on travis are all passing too. Check out [Latest Test](https://travis-ci.org/github/amanchokshi/EMBERS/jobs/744106700) and [All Past Tests](https://travis-ci.org/github/amanchokshi/EMBERS/builds)

I'm stumped too, i'm going to ignore this for the review, as the examples seem to run (ok , i do have one failure still, the ephem_chrono task).
