### 2020-11-19 #2629 @amanchokshi
Hi Peter, I'll try and summarise what I've found in my performance tests:

- O(N) - Performance scales linearly with the number of files, on a single core
- O(M) - Performance scales linearly with the number of lines of data, with a significant overhead reading in the data. By my estimate, more than half the time processing a single file of ~17k lines can be attributed to reading in the file. This is ~0.27 seconds per file. The processing itself is done with `scipy` interpolation + filtering, which scales linearly with array size
