### 2020-11-19 #2629 @teuben
> thanks Aman, this is indeed very useful stuff to understand.  50% in the I/O is quite a lot.  These are 2kb files, that should read in 0.02ms on a typical 100MB/s disk. So a significant time it does something else. Maybe the decoding?  Even the readlines() routine isn't able to perform at formal speed. I measured indeed about 250ms for reading one file. When not processing any line, so returning nothing, it still look 5-10 ms. I might play with this, I didn't see any obvious ways to speed this up.

oops, i messed up my units.  The files are 2MB, so that should be around 10ms, which corresponds to my just processing the lines, no decoding.   So that decoding is expensive, more like 200 ms, 20 times as expensive as reading the line.
