### 2020-11-19 d0f935d
Moving to version 0.8.2 to fix the default option of --max_cores on batch scripts
