### 2020-11-19 3b4e3bd
Fixed the default option of --max_cores on batch scripts
