+++ src/embers/kindle/rfe_calibration.py
@@ -139,7 +139,6 @@ def main():
     _parser.add_argument(
         "--max_cores",
         metavar="\b",
-        default=0,
         type=int,
         help="Maximum number of cores to be used by this script. By default all core available cores are used",
     )
@@ -161,9 +160,6 @@ def main():
     _chan_map_dir = _args.chan_map_dir
     _out_dir = _args.out_dir
     _max_cores = _args.max_cores
-
-    if _max_cores == 0:
-        _max_cores is None
 
     print(f"RF Explorer calibration files saved to: {_out_dir}")
     rfe_batch_cali(
