+++ src/embers/kindle/ephem_batch.py
@@ -62,7 +62,6 @@ def main():
     _parser.add_argument(
         "--max_cores",
         metavar="\b",
-        default=0,
         type=int,
         help="Maximum number of cores to be used by this script. By default all core available cores are used",
     )
@@ -74,9 +73,6 @@ def main():
     _alpha = _args.alpha
     _out_dir = _args.out_dir
     _max_cores = _args.max_cores
-
-    if _max_cores == 0:
-        _max_cores is None
 
     print(f"Saving logs to {_out_dir}/ephem_data")
     print(f"Saving sky coverage plots to {_out_dir}/ephem_plots")
