+++ src/embers/kindle/tile_maps.py
@@ -143,7 +143,6 @@ def main():
     _parser.add_argument(
         "--max_cores",
         metavar="\b",
-        default=0,
         type=int,
         help="Maximum number of cores to be used by this script. By default all core available cores are used",
     )
@@ -166,9 +165,6 @@ def main():
     _plots = _args.plots
     _rfe_cali_bool = _args.rfe_cali_bool
     _max_cores = _args.max_cores
-
-    if _max_cores == 0:
-        _max_cores is None
 
     if _plots == "False":
         _plots is False
