+++ src/embers/kindle/sat_channels.py
@@ -97,7 +97,6 @@ def main():
     _parser.add_argument(
         "--max_cores",
         metavar="\b",
-        default=0,
         type=int,
         help="Maximum number of cores to be used by this script. By default all core available cores are used",
     )
@@ -114,9 +113,6 @@ def main():
     _out_dir = _args.out_dir
     _plots = _args.plots
     _max_cores = _args.max_cores
-
-    if _max_cores == 0:
-        _max_cores is None
 
     if _plots == "True":
         _plots = True
