### 2020-11-19 #2629 @amanchokshi
The batch scripts should now work after you update EMBERS to the latest version - 0.8.2. I had added a bad default value to the max_cores argparse option.
