### 2020-11-19 #2629 @amanchokshi
I ran `ephem_chrono` again and things worked. Do you have a folder called `TLE` in the `sat_utils` directory? This should have been copied over when you cloned the sample data and moved the sample data to the correct dir.

Before running `ephem_batch`:

```
tree -d sat_utils
sat_utils
└── TLE
```

After running `ephem_batch`:

```
tree -d sat_utils
sat_utils
├── TLE
├── ephem_data
└── ephem_plots
```
The above data is required for `ephem_chrono` to work. Finally, after `ephem_chrono` runs successfully:

```
tree -d sat_utils
sat_utils
├── TLE
├── ephem_chrono
├── ephem_data
└── ephem_plots
```

Can you delete all the data in the `sat_utils` directory, except the `TLE` folder, then run `ephem_batch` followed by `ephem_chrono`?
