### 2020-11-19 8bb5b92
Added a section of performance to the documentation
