### 2020-11-19 #2629 @amanchokshi
I've created a new [Performance](https://embers.readthedocs.io/en/latest/performance.html) page in the documentation. This outlines our investigation into parallel and single core performance and  states the I/O limitations. I've also added a link to the performance page in the examples.
