### 2020-11-19 #2629 @teuben
I've removed all directories (except TLE) in embers_out/sat_utils, the TLE directory has the following 73 files:

21576.txt  25113.txt  25117.txt  25159.txt  25415.txt  25419.txt  25477.txt  25481.txt  25982.txt  25986.txt  33062.txt  40069.txt  40089.txt  41180.txt  41184.txt  41188.txt  44020.txt  44025.txt  44387.txt
23545.txt  25114.txt  25118.txt  25338.txt  25416.txt  25420.txt  25478.txt  25482.txt  25983.txt  28654.txt  33063.txt  40086.txt  40090.txt  41181.txt  41185.txt  41189.txt  44022.txt  44026.txt
23546.txt  25115.txt  25119.txt  25413.txt  25417.txt  25475.txt  25479.txt  25980.txt  25984.txt  33060.txt  33064.txt  40087.txt  40091.txt  41182.txt  41186.txt  44018.txt  44023.txt  44027.txt
25112.txt  25116.txt  25158.txt  25414.txt  25418.txt  25476.txt  25480.txt  25981.txt  25985.txt  33061.txt  33065.txt  40088.txt  41179.txt  41183.txt  41187.txt  44019.txt  44024.txt  44028.txt

Running ephem_chrono immediately returns (though it has the coffee message)

$ ephem_chrono
Saving chronological Ephem files to: ./embers_out/sat_utils/ephem_chrono
Grab a coffee, this may take more than a couple of minutes!

It does create all the json files, but they just contain [], and the log file has 0 lines.

I am doing this in a clean directory, where i only see embers_out and tiles_data.   The code is somewhere else, which I installed using "pip install -e ."    I am using a "venv", which is a directory inside the EMBERS source tree.
