### 2020-11-19 #2629 @amanchokshi
This is bizarre. Let try a couple of long shots. Can you check whether the files in the TLE directory contain anything? Perhaps they've been corrupted somehow?

You should see something like this:

```
$ tail embers_out/sat_utils/TLE/21576.txt

1 21576U 91050C   19282.72964883  .00000014  00000-0  18223-4 0  9990
2 21576  98.6203 331.5014 0002898 326.0593  34.0410 14.41572163482935
1 21576U 91050C   19283.90959667 +.00000017 +00000-0 +19271-4 0  9992
2 21576 098.6201 332.6919 0002839 321.6987 038.4000 14.41572293482883
```

Have you been running all the other examples in the same way? With EMBERS installed using `pip install -e .` in the source tree?

I've attempted to replicate the way you're running things. I have my source directory at `~/Repos/EMBERS` within which I've created a virtual environment and run `pip install --editable .` If I do a pip list when this virtual environment is active, I see that `embers` version `0.8.2` is installed form source

```
Package                       Version     Location
----------------------------- ----------- -----------------------------------
embers                        0.8.2       /Users/amanchokshi/Repos/EMBERS/src
```
I've then created a new directory `~/Repos/em_test/` which just contains `tiles_data` and `embers_out/sat_utils/TLE`

Keeping the original virtual env active, I run the following

```
$ ephem_batch
$ ephem_chrono
```

Checking the contents of the first file in the `ephem_chrono` dir:

```
$ tail embers_out/sat_utils/ephem_chrono/2019-10-10-00:00.json

            1.5377585726597032,
            1.5366094819324243,
            1.5354642595022225,
            1.534322886123595,
            1.5331853425510376,
            1.5320516095390462,
            1.5309216678421183
        ]
    }
]%
```

Have I deviated from your setup in any way?

PS: It does usually take a while to process the data (~10 mins), if we were working with a larger data set. The sample data only contains a couple of satellite passes each. I'll remove the the coffee message in the next EMBERS version.
