### 2020-11-19 afde898
Explicitely close plots in save_ephem function and remove coffee statement
