### 2020-11-19 #2629 @amanchokshi
So many `matplotlib` figures should not be opened by `ephem_batch`, instead figures should be saved to `embers_out/sat_utils/ephem_plots`. Does `ephem_batch` create this directory and `.png` figures for you?

Can you update EMBERS to V0.8.3, or `git pull` and build from source again? I've explicitly closed the figures after the plot is saved. This shouldn't be necessary, but may help.

Also have removed the coffee message, which was unnecessary
