### 2020-11-19 #2629 @teuben
Indeed my embers_out/sat_utils/TLE/21576.txt looks like yours. I've been able to run all steps before this fatal one just fine, all the figure are like yours.

You have also done what I did , with the pip installing it with source in place.   I use a pretty standard anaconda3 as the base for my python, not the system python. I've actually reinstalled that one.

I will retrace the steps, but a question before I do this, can i reinstall and just do ephem_batch, it doesn't need  the lengthy align_batch, does it ?

A suggestion on your side might be to have some kind of -v flag for all your scripts, which makes it verbose, to perhaps help to see some expected intermediate output, helping you to see what's going on here. In my own system I have debug=LEVEL, which will create more and more output the higher the level is.    Other packages have -vvv instead of -v to increase the verbosity level :[)   Even for myself I find it useful, as I forget my own code, and a debug level could help with that.
