### 2020-11-19 #2629 @teuben
> So many `matplotlib` figures should not be opened by `ephem_batch`, instead figures should be saved to `embers_out/sat_utils/ephem_plots`. Does `ephem_batch` create this directory and `.png` figures for you?
>
> Can you update EMBERS to V0.8.3, or `git pull` and build from source again? I've explicitly closed the figures after the plot is saved. This shouldn't be necessary, but may help.
>
> Also have removed the coffee message, which was unnecessary

I noted that 3 files in TLE have 3 length:44019.txt  44020.txt  44025.txt - the rest have anywhere from 20 to 78 lines. There are 73 files in here.

I do see 70 plots in the ephem_plots directory. (not 73, since 3 of my TLE's were 0 length)

I've updated code again. Ran ephem_chrone, but still no json files.
