### 2020-11-19 858634c
Added warning to use --max_cores=1 on laptops
