### 2020-11-19 #2629 @amanchokshi
Just a thought, is the directory with `tiles_data` and `embers_out/sat_utils/TLE` on your external `VFAT` formatted drive? The names of the json files created have a colon `:` in them. Could that be tripping things up?
