### 2020-11-19 #2629 @teuben
> You don't need to run `align_batch` again. The only thing that any of the `ephem` scripts should need are the TLE files. It's also okay that 3 of the TLE files are empty, they have very sparse satellite coverage and on that data had np passes visible from the site.
>
> I've been using `pyenv` to install python versions on my system. Currently I've used `pyenv` to set by default version to `3.8.3`. I used `pyenv virtualenv` to create my virtual enviroments. I don't think this should be an issue.
>
> Are you on a python version >= 3.6? I have used f-strings, do 3.5 will have problems

python 3.8.5
the whole tree with embers_out and tiles_data is on a regular ext4 formatted drive.
The json files look fine in their name, e.g.
       -rw-rw-r-- 1 teuben teuben 2 Nov 19 09:00 embers_out/sat_utils/ephem_chrono/2019-10-10-00:00.json
