### 2020-11-19 #2629 @amanchokshi
Lets rule out problems in `ephem_batch` so that we can focus on `ephem_chrono`. Can you try and see if the log seems similar?

```
$ tail embers_out/sat_utils/ephem_data/ephem_batch.log

INFO: ephem_batch: Saved sky coverage plot of satellite [25481] to ./embers_out/sat_utilsephem_plots/25481.png
Saved ephemeris of satellite [25481] to ./embers_out/sat_utilsephem_data/25481.npz
INFO: ephem_batch: Saved sky coverage plot of satellite [23546] to ./embers_out/sat_utilsephem_plots/23546.png
Saved ephemeris of satellite [23546] to ./embers_out/sat_utilsephem_data/23546.npz
INFO: ephem_batch: Saved sky coverage plot of satellite [25479] to ./embers_out/sat_utilsephem_plots/25479.png
Saved ephemeris of satellite [25479] to ./embers_out/sat_utilsephem_data/25479.npz
INFO: ephem_batch: Saved sky coverage plot of satellite [23545] to ./embers_out/sat_utilsephem_plots/23545.png
Saved ephemeris of satellite [23545] to ./embers_out/sat_utilsephem_data/23545.npz
INFO: ephem_batch: Saved sky coverage plot of satellite [25478] to ./embers_out/sat_utilsephem_plots/25478.png
Saved ephemeris of satellite [25478] to ./embers_out/sat_utilsephem_data/25478.npz
```

Also that the first plot `embers_out/sat_utils/ephem_plots/21576.png` looks like this:

![21576](https://user-images.githubusercontent.com/40157844/99679853-881b8080-2ad0-11eb-889c-1b4341201360.png)

And finally that the size of the `ephem_data` directory matches mine:

```
$ du -sh embers_out/sat_utils/ephem_data

>>> 2.2M    embers_out/sat_utils/ephem_data
```
