### 2020-11-19 #2629 @teuben
log file looks ok, but has a different order of the files downloaded.
File has 143 lines (73 for files, 70 for successfull ones)

the plot looks similar, though mine has 41 passes, yours has 5.

my du -sh embers_out/sat_utils/ephem_data
17M     embers_out/sat_utils/ephem_data

e.g. the one you plotted, where I had 41 passes:

-rw-rw-r-- 1 teuben teuben 146855 Nov 19 08:26 21576.npz

It has 70 files (3 missing due to 0 length TLE files)
