### 2020-11-19 #2629 @amanchokshi
It's really strange that you have 41 satellite passes. Can you try this

```
$ ephem_single --sat=21576 --out_dir=tmp

$ tree tmp

tmp
├── ephem_data
│   └── 21576.npz
└── ephem_plots
    └── 21576.png
```
The `.npz` data file should be only `21K` in size. If the plot has more than 5 satellites, it means that something is wrong with the TLE files. This would explain why `ephem_chrono` is making empty files
