### 2020-11-19 #2629 @amanchokshi
It's a long shot, but can you check if the contents of the TLE file is the same as mine:

```
$ more embers_out/sat_utils/TLE/21576.txt

1 21576U 91050C   19282.72964883  .00000014  00000-0  18223-4 0  9990
2 21576  98.6203 331.5014 0002898 326.0593  34.0410 14.41572163482935
1 21576U 91050C   19283.90959667 +.00000017 +00000-0 +19271-4 0  9992
2 21576 098.6201 332.6919 0002839 321.6987 038.4000 14.41572293482883

```
