### 2020-11-19 #2629 @teuben
ls -l tmp/ephem_*/*
-rw-rw-r-- 1 teuben teuben 146855 Nov 19 10:18 tmp/ephem_data/21576.npz
-rw-rw-r-- 1 teuben teuben 171585 Nov 19 10:18 tmp/ephem_plots/21576.png

so much bigger.   The png shows 41 passes.
