### 2020-11-19 #2629 @teuben
mine is clearly longer

more embers_out/sat_utils/TLE/21576.txt
1 21576U 91050C   19274.12296826  .00000022  00000-0  20647-4 0  9992
2 21576  98.6215 322.8165 0003370 355.9403   4.1757 14.41571172481698
1 21576U 91050C   19274.74764681 +.00000022 +00000-0 +20705-4 0  9994
2 21576 098.6214 323.4469 0003330 354.0023 006.1125 14.41571283481565
1 21576U 91050C   19275.64996028  .00000018  00000-0 19615-4 0  9999
2 21576  98.6213 324.3575 0003275 351.1025   9.0105 14.41571404481914
1 21576U 91050C   19276.69109109  .00000011  00000-0 17383-4 0  9994
2 21576  98.6211 325.4081 0003216 347.5805  12.5303 14.41571520482061
1 21576U 91050C   19277.73222188  .00000003  00000-0 14745-4 0  9993
2 21576  98.6209 326.4587 0003154 343.9674  16.1414 14.41571600482214
1 21576U 91050C   19278.70394390 -.00000000 +00000-0 +13856-4 0  9996
2 21576 098.6208 327.4393 0003103 340.6734 019.4337 14.41571694482138
1 21576U 91050C   19279.12039615  .00000001  00000-0 14345-4 0  9993
2 21576  98.6207 327.8595 0003082 339.1436  20.9627 14.41571750482416
1 21576U 91050C   19279.74507450 +.00000006 +00000-0 +15778-4 0  9994
2 21576 098.6207 328.4898 0003052 336.8952 023.2099 14.41571851482289
1 21576U 91050C   19280.36975290  .00000009  00000-0 16766-4 0  9996
2 21576  98.6206 329.1202 0003016 334.7424  25.3617 14.41571918482593
1 21576U 91050C   19280.71679643 +.00000012 +00000-0 +17538-4 0  9990
2 21576 098.6206 329.4703 0002999 333.5040 026.5995 14.41571968482421
1 21576U 91050C   19281.68851827  .00000014  00000-0 18394-4 0  9996
2 21576  98.6205 330.4508 0002950 329.8875  30.2144 14.41572084482783
1 21576U 91050C   19282.72964883  .00000014  00000-0 18223-4 0  9990
2 21576  98.6203 331.5014 0002898 326.0593  34.0410 14.41572163482935
