### 2020-11-19 #2629 @amanchokshi
The data you have doesn't match what's on github currently. I initially had too much data on the sample data repo and had to reduce it to make the examples run in a reasonable time.

Can you pull down the sample data repo again, or just do download the TLE dir from the [sample data repo](https://github.com/amanchokshi/EMBERS-Sample-Data)?

The new TLE files should be sufficient to get `ephem_chrono` to work.

Can you check whether all the files in `tiles_data` have the time-stamp `2019-10-10` in them? If they do, you have the correct data and the wrong TLE files. If they don't you'll unfortunately have do pull the whole sample data repo again.

Sorry for the hassle. Hopefully it's smooth sailing from here 🤞
