### 2020-11-20 #2629 @teuben
I redid the data repo instructions. I now have a short embers_out/sat_utils/TLE/21576.txt file, with 4 lines.
The tiles_data are all (1440 I see) from 2019-10-10

I'm afraid, even with a new  align_batch and ephem_batch  I still get a blank ephem_chrono.

I think I'm going to redo the whole workflow.
