### 2020-11-20 #2629 @amanchokshi
Maybe you could make a bash script with just the important stuff and let it run?

```
#!/bin/bash

align_batch --max_cores=1
ephem_batch --max_cores=1
ephem_chrono
```
