### 2020-11-20 #2629 @teuben
Although I had done exactly those steps, now it's producing non-zero json files. I see 48 json files, of which the first 16 have a good size. Until 2019-10-10-07:30   The ones from 2019-10-10-08:00.json to 2019-10-10-23:30.json are blanked.
