### 2020-11-20 #2629 @amanchokshi
Hi Peter, the last files which are empty are the result of our small sample data set. I'm appending a screenshot of all `ephem_chrono` json files produced when I use the full data set.

![Screen Shot 2020-11-21 at 10 44 21 am](https://user-images.githubusercontent.com/40157844/99860212-8643e100-2be6-11eb-831c-dd092918e0b0.png)
