### 2020-11-24 #2629 @mbobra
@amanchokshi We're almost there! Can you please [archive your release on Zenodo](https://guides.github.com/activities/citable-code/) to obtain a DOI and then put that in your README.rst file? Please make sure the authors listed on the Zenodo entry match those listed on the JOSS paper. After that I think we're done 🎉
