### 2020-11-24 db0dfa4
Moving to v1.0.0 for JOSS and Zenodo release
