### 2020-11-24 #2629 @amanchokshi
@mbobra Thanks Monica! I've uploaded EMBERS to Zenodo and have added the DOI to the README.rst :)
