### 2020-11-24 #2629 @mbobra
@openjournals/joss-eics This paper is ready for final processing 🎉 Congratulations @amanchokshi!

Thank you very much for your thorough review, @teuben. Your contributions keep JOSS running. I really appreciate all your hard work ☀️
