### 2020-11-25 #2629 @arfon
@teuben, @mbobra - many thanks for your reviews here and @mbobra for editing too! JOSS simply wouldn't work without the volunteer efforts of folks like yourselves ✨
